% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clocaenog}
\alias{clocaenog}
\title{Clocaenog 6 sample data}
\format{
An object of class \code{list} of length 2.
}
\usage{
clocaenog
}
\description{
Sample data of tree species and their position that have been sampled by Arne
Pommerening, 2012 (https://www.pommerening.org).
}
\details{
The list contains two items, first the tree specific data, seconda look-up
table of the species names.
}
\keyword{datasets}
