% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{depthList}
\alias{depthList}
\title{Determine depth of a list}
\usage{
depthList(list)
}
\arguments{
\item{list}{[\code{list(.)}] list to test.}
}
\value{
an integer value of the depth of a list.
}
\description{
Determine depth of a list
}
\examples{
x <- list(int = c(1:5),
          char = list(lower = c(letters[1:5]),
                      upper = c(LETTERS[1:5])))
depthList(x)
depthList(x[[1]])
}
