% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getDecimals}
\alias{.getDecimals}
\title{Get the number of decimal places}
\usage{
.getDecimals(x)
}
\arguments{
\item{x}{[\code{numeric(1)}]\cr the number for which to derive decimal
places.}
}
\description{
Get the number of decimal places
}
