% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.makeColours}
\alias{.makeColours}
\title{Gather colours for the plot}
\usage{
.makeColours(input = NULL, theme = NULL, ...)
}
\arguments{
\item{input}{[]}

\item{theme}{[]}

\item{...}{[]}
}
\description{
Gather colours for the plot
}
