% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLayout.R
\name{.makeLayout}
\alias{.makeLayout}
\title{Make the layout of a plot}
\usage{
.makeLayout(legend = NULL, window = NULL, theme = gtTheme)
}
\arguments{
\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}

\item{x}{[\code{list(.)}]\cr the object, output from \code{makeObject}, from
which to make the plot.}
}
\description{
Make the layout of a plot
}
