% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_.R
\name{.makePlot}
\alias{.makePlot}
\title{Make the object to a plot}
\usage{
.makePlot(x, window, theme = geoTheme, ...)
}
\arguments{
\item{x}{\link[=list]{list(1)}\cr named list of the object from which to make the
plot.}

\item{window}{\link[=data.frame]{data.frame(2)}\cr in case the reference window
deviates from the bounding box of \code{crds}, specify here the minimum and
maximum values in columns \code{x} and \code{y}.}

\item{theme}{\link[=gtTheme]{geometr theme}\cr the theme from which to take
graphical parameters.}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}
}
\description{
Make the object to a plot
}
