% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_.R
\name{.makeTinyMap}
\alias{.makeTinyMap}
\title{Make a tiny map}
\usage{
.makeTinyMap(geom = NULL)
}
\arguments{
\item{geom}{\link[=geom]{gridded(1)}\cr the geom from which to create a tiny map.}
}
\description{
A tiny map is used via the show method of a geom.
}
