% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testAnchor}
\alias{.testAnchor}
\title{Test anchor for consistency}
\usage{
.testAnchor(x, ...)
}
\arguments{
\item{x}{[\code{data.frame | geom}]\cr the object to be tested for
consistency.}

\item{...}{[\code{.}]\cr additional arguments.}
}
\description{
Test anchor for consistency
}
