% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testColours}
\alias{.testColours}
\title{Test colours}
\usage{
.testColours(colours = NULL)
}
\arguments{
\item{colours}{[\code{.}]\cr the items to test for whether they are a colour
value that is valid in R.}
}
\description{
Test colours
}
\details{
inspired by \href{https://stackoverflow.com/a/13290832}{https://stackoverflow.com/a/13290832}
}
