% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testPoints}
\alias{.testPoints}
\title{Test points to not contain NA}
\usage{
.testPoints(x, ...)
}
\arguments{
\item{x}{\link[=data.frame]{data.frame(2)}\cr the points to be tested.}

\item{...}{additional arguments.}
}
\description{
Test points to not contain NA
}
