% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tester.R
\name{.testTemplate}
\alias{.testTemplate}
\title{Test template for consistency}
\usage{
.testTemplate(x, ...)
}
\arguments{
\item{x}{[gridded(1)][geom]\cr the object to be tested for
consistency.}

\item{...}{\cr additional arguments.}
}
\description{
Test template for consistency
}
