% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testWindow}
\alias{.testWindow}
\title{Test window for consistency}
\usage{
.testWindow(x, ...)
}
\arguments{
\item{x}{\link[=data.frame]{data.frame(2)}\cr the minimum and maximum values in
columns \code{x} and \code{y}.}

\item{...}{additional arguments.}
}
\description{
Test window for consistency
}
