% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTheme.R
\name{.updateTheme}
\alias{.updateTheme}
\title{Update the theme parameters}
\usage{
.updateTheme(x, theme = gtTheme)
}
\arguments{
\item{x}{[\code{list(1)}]\cr any spatial object to plot.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme that shall be updated.}
}
\description{
Update the theme parameters
}
