% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateWindow}
\alias{.updateWindow}
\title{Update the window slot}
\usage{
.updateWindow(geom = NULL, window = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr a geom for which a new window should be derived.}

\item{window}{[\code{tibble(1)}]\cr the old window.}
}
\description{
Set the extent of a window to smaller/larger values, if the vertices would be
beyond the window otherwise.
}
