% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_grob.R
\docType{methods}
\name{gc_grob}
\alias{gc_grob}
\alias{gc_grob,geom-method}
\alias{gc_grob,sf-method}
\alias{gc_grob,Spatial-method}
\title{Make a geometry object of class grob}
\usage{
\S4method{gc_grob}{geom}(input, theme = gtTheme, ...)

\S4method{gc_grob}{sf}(input, theme = gtTheme, ...)

\S4method{gc_grob}{Spatial}(input, theme = gtTheme, ...)
}
\arguments{
\item{input}{the object from which to make an object of class \code{grob}.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to make parameters.}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  \code{\link{polylineGrob}} or a \code{\link{pathGrob}}.
}
\description{
Make a geometry object of class grob
}
\seealso{
Other spatial classes: \code{\link{gc_geom}},
  \code{\link{gc_sf}}, \code{\link{gc_sp}}
}
\concept{spatial classes}
