% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_ppp.R
\name{gc_ppp}
\alias{gc_ppp}
\alias{gc_ppp,geom-method}
\title{Transform a spatial object to class \code{ppp}}
\usage{
\S4method{gc_ppp}{geom}(input = NULL)
}
\arguments{
\item{input}{the object to transform to class \code{ppp}.}
}
\value{
an object of class \code{ppp}
}
\description{
Transform a spatial object to class \code{ppp}
}
\examples{
gc_ppp(input = gtGeoms$point)
}
\seealso{
Other spatial classes: 
\code{\link{gc_geom}()},
\code{\link{gc_grob}()},
\code{\link{gc_raster}()},
\code{\link{gc_sf}()},
\code{\link{gc_sp}()}
}
\concept{spatial classes}
