% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_raster.R
\name{gc_raster}
\alias{gc_raster}
\alias{gc_raster,geom-method}
\title{Transform a spatial object to class \code{Raster*}}
\usage{
gc_raster(input)

\S4method{gc_raster}{geom}(input = NULL)
}
\arguments{
\item{input}{the object to transform to class \code{Raster*}.}
}
\value{
an object of class \code{Raster*}
}
\description{
Transform a spatial object to class \code{Raster*}
}
\examples{
gc_raster(input = gtGeoms$grid$categorical)
}
\seealso{
Other spatial classes: 
\code{\link{gc_geom}()},
\code{\link{gc_sf}()},
\code{\link{gc_sp}()}
}
\concept{spatial classes}
