% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_sp.R
\docType{methods}
\name{gc_sp}
\alias{gc_sp}
\alias{gc_sp,geom-method}
\title{Transform a spatial object to class \code{sp}}
\usage{
\S4method{gc_sp}{geom}(input = NULL)
}
\arguments{
\item{input}{the object to transform to class \code{sp}.}
}
\value{
an object of class \code{sp}
}
\description{
Transform a spatial object to class \code{sp}
}
\examples{
(spPoints <- gc_sp(input = gtGeoms$point))
(spLines <- gc_sp(input = gtGeoms$line))
(spPolygon <- gc_sp(input = gtGeoms$polygon))
}
\seealso{
Other spatial classes: \code{\link{gc_geom}},
  \code{\link{gc_grob}}, \code{\link{gc_ppp}},
  \code{\link{gc_sf}}
}
\concept{spatial classes}
