% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_terra.R
\name{gc_terra}
\alias{gc_terra}
\alias{gc_terra,geom-method}
\title{Transform a spatial object to class \code{SpatRaster}}
\usage{
gc_terra(input)

\S4method{gc_terra}{geom}(input = NULL)
}
\arguments{
\item{input}{the object to transform to class \code{SpatRaster}.}
}
\value{
an object of class \code{SpatRaster}
}
\description{
Transform a spatial object to class \code{SpatRaster}
}
\examples{
gc_terra(input = gtGeoms$grid$categorical)
}
\seealso{
Other spatial classes: 
\code{\link{gc_geom}()},
\code{\link{gc_raster}()},
\code{\link{gc_sf}()},
\code{\link{gc_sp}()}
}
\concept{spatial classes}
