% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_filter.R
\name{geo_filter}
\alias{geo_filter}
\title{Subset a geometric object using column values}
\usage{
geo_filter(obj, ..., update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to derive a subset
from.}

\item{...}{subset based on logical predicates defined in terms of the columns
in \code{x} (of both, points, features and groups). Multiple conditions are
combined with \code{&}. Only rows where the condition evaluates to TRUE are
kept.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
after deriving the subset.}
}
\value{
\code{geom} of the subset of \code{obj}.
}
\description{
This function allows to subset any geometric object for which all required
getters are available.
}
\examples{
geo_filter(obj = gtGeoms$point, y < -10)

geo_filter(obj = gtGeoms$point, attr \%in\% c("A", "C", "E"))
}
\seealso{
Other geometry tools: 
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_skew}()},
\code{\link{geo_stretch}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
