% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_pull.R
\name{geo_pull}
\alias{geo_pull}
\title{Extract a single column from a geometric object}
\usage{
geo_pull(obj, var, ungroup = FALSE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to pull a column from.}

\item{var}{\link[=character]{character(1)}\cr name of the variable to pull.}

\item{ungroup}{\link[=logical]{logical(1)}\cr this argument provides the attribute to
pull per each individual feature, producing duplicates in case there is
more than one feature per group.}
}
\value{
vector of the column specified in \code{var}.
}
\description{
This function allows to extract a specific column from any geometric object
for which all required getters are available and thus reflects the base
function \code{$}.
}
\details{
This function searches for \code{var} by first looking in the
groups, then the features and finally the points of \code{obj}. This
results always in an output that is limited to the unique cases of
\code{var}. In case you want the explicit values of, for instance,
\code{fid} in \code{obj@points}, you have to extract points and then use
\code{\link[dplyr]{pull}} on the result.
}
\examples{
# pull values from a geom (there are two features, thus two values) ...
geo_pull(obj = gtGeoms$point, var = "fid")

# pull from a Raster* with RAT
geo_pull(obj = gtGeoms$grid, var = "cover")
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_skew}()},
\code{\link{geo_stretch}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
