% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_random.R
\name{geo_random}
\alias{geo_random}
\title{Create a \code{geom} randomly}
\usage{
geo_random(type = "point", window = NULL, vertices = NULL)
}
\arguments{
\item{type}{\link[=character]{character(1)}\cr Either one of the three main feature
types \code{"point"}, \code{"line"} or \code{"polygon"}, \code{"random"},
or more specifically one of their subtypes, e.g. \code{"hexagon"} (subtypes
currently not yet supported).}

\item{window}{\link[=data.frame]{data.frame(2)}\cr in case the reference window
deviates from the bounding box of \code{crds}, specify here the minimum and
maximum values in columns \code{x} and \code{y}.}

\item{vertices}{\link[=integer]{integerish(1)}\cr the number of vertices the geometry
should have; only meaningful if \code{type} does not indicate the number of
vertices already. If left at \code{NULL} the minimum number of vertices for
the \code{geom} type, i.e. 1 for \code{point}, 2 for \code{line} and 3 for
\code{polygon}.}
}
\value{
A \code{geom}.
}
\description{
This function creates a random geometry
}
\examples{
# create a random polygon with five vertices
set.seed(1)
someGeom <- geo_random(type = "polygon", vertices = 5)
geo_vis(geom = someGeom, linecol = "#FFB000")

# in case template is given, this serves as source for the window extent
geo_vis(geom = geo_random(), linecol = "#B24223", pointsymbol = 22, new = FALSE)
}
\seealso{
Other geometry shapes: 
\code{\link{geo_line}()},
\code{\link{geo_point}()},
\code{\link{geo_polygon}()}
}
\concept{geometry shapes}
