% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_reflect.R
\name{geo_reflect}
\alias{geo_reflect}
\title{Reflect geometric objects}
\usage{
geo_reflect(obj, angle, fid = NULL, update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to reflect.}

\item{angle}{\link[=numeric]{numeric(1)}\cr the counter-clockwise angle by which the
reflection axis shall be rotated (can be negative to rotate clockwise).}

\item{fid}{\link[=integer]{integerish(1)}\cr in case only a subset of features shall
be rotated, specify that here.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
after rotation.}
}
\value{
\code{geom} of the reflected \code{obj}.
}
\description{
Reflect geometric objects across a reflection axis.
}
\details{
The reflection axis is a straight line that goes through the plot
origin with the given angle, where positive angles open towards the
positive y-axis and negative angles open up towards the negative y-axis.
}
\examples{
# reflect several features
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_reflect(obj = gtGeoms$polygon, angle = 45,
                       update = FALSE)
geo_vis(newPoly, linecol = "green", new = FALSE)

# reflect a single feature
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_reflect(obj = gtGeoms$polygon, angle = 90, fid = 2,
                       update = FALSE)
geo_vis(newPoly, linecol = "green", new = FALSE)
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_skew}()},
\code{\link{geo_stretch}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
