% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_scale.R
\name{geo_scale}
\alias{geo_scale}
\title{Scale geometric objects}
\usage{
geo_scale(obj, range = NULL, fid = NULL, update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to be scaled.}

\item{range}{\link[=data.frame]{data.frame(2)}\cr vector of length two for both of the
\code{x} and \code{y} dimension to which the values should be scaled.}

\item{fid}{\link[=integer]{integerish(.)}\cr in case only a subset of features shall
be scaled, specify that here.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
of the resulting geom.}
}
\value{
\code{geom} of the scaled \code{obj}.
}
\description{
Scale the vertex values of geometric objects to a values range.
}
\examples{
# rescale to values between -10 and 10
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_scale(obj = gtGeoms$polygon, update = FALSE,
                     range = data.frame(x = c(-10, 10), y = c(-10, 10)))
geo_vis(geom = newPoly, linecol = "green", new = FALSE)

# rescale a single feature
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_scale(obj = gtGeoms$polygon, update = FALSE, fid = 2,
                     range = data.frame(x = c(-10, 10), y = c(-10, 10)))
geo_vis(geom = newPoly, linecol = "green", new = FALSE)
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_skew}()},
\code{\link{geo_stretch}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
