% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_skew.R
\name{geo_skew}
\alias{geo_skew}
\title{Skew geometric objects}
\usage{
geo_skew(obj, x = NULL, y = NULL, fid = NULL, update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to skew.}

\item{x}{\link[=numeric]{numeric(1)}\cr the shear factor in x dimension.}

\item{y}{\link[=numeric]{numeric(1)}\cr the shear factor in y dimension.}

\item{fid}{\link[=integer]{integerish(.)}\cr in case only a subset of features shall
be skewed, specify that here.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
after skewing.}
}
\value{
\code{geom} of the skewed \code{obj}.
}
\description{
Skew geometric objects by a shear factor in x and y dimension.
}
\examples{
# skew several features
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_skew(obj = gtGeoms$polygon, x = 0.5, update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)

# skew a single feature
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_skew(obj = gtGeoms$polygon, x = 0.5, y = .7, fid = 2,
                   update = FALSE)
geo_vis(newPoly, linecol = "green", new = FALSE)
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_stretch}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
