% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_stations}
\alias{geo_stations}
\title{Query Water survey of Canada statiosn}
\usage{
geo_stations(STATION_NUMBER = NULL, as_spatial = FALSE)
}
\arguments{
\item{STATION_NUMBER}{7 digit unique station ID}

\item{as_spatial}{Convert to sf object}
}
\description{
Query geospatial webservice for hydrometric data
}
