% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_stretch.R
\name{geo_stretch}
\alias{geo_stretch}
\title{Stretch geometric objects}
\usage{
geo_stretch(obj, x = NULL, y = NULL, fid = NULL, update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to stretch.}

\item{x}{\link[=numeric]{numeric(1)}\cr the scale factor in x dimension.}

\item{y}{\link[=numeric]{numeric(1)}\cr the scale factor in y dimension.}

\item{fid}{\link[=integer]{integerish(.)}\cr in case only a subset of features shall
be stretched, specify that here.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
after stretching.}
}
\value{
\code{geom} of the stretched \code{obj}.
}
\description{
Stretch geometric objects by a scale factor in x and y dimension.
}
\examples{
# stretch several features
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_stretch(obj = gtGeoms$polygon, x = 0.5, y = 0.2,
                       update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)

# stretch a single feature
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_stretch(obj = gtGeoms$polygon, x = 0.5, fid = 2, update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)

# stretch feature separately
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_stretch(obj = gtGeoms$polygon,
                      x = c(0.2, 1),
                      y = c(1, 0.2),
                      update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_skew}()},
\code{\link{geo_translate}()}
}
\concept{geometry tools}
