% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_translate.R
\name{geo_translate}
\alias{geo_translate}
\title{Translate geometric objects}
\usage{
geo_translate(obj, x = NULL, y = NULL, fid = NULL, update = TRUE)
}
\arguments{
\item{obj}{\link[=geom]{gridded(1)}\cr the object to translate.}

\item{x}{\link[=numeric]{numeric(1)}\cr the translation constant (offset) in x
dimension.}

\item{y}{\link[=numeric]{numeric(1)}\cr the translation constant (offset) in y
dimension.}

\item{fid}{\link[=integer]{integerish(.)}\cr in case only a subset of features shall
be translated, specify that here.}

\item{update}{\link[=logical]{logical(1)}\cr whether or not to update the window slot
after translation.}
}
\value{
\code{geom} of the mathematically translated \code{obj}.
}
\description{
Translate geometric objects by adding a constant in x and y dimension.
}
\examples{
# translate several features
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_translate(obj = gtGeoms$polygon, x = 5, y = c(-10, 5),
                        update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)

# translate a single feature
geo_vis(gtGeoms$polygon, linewidth = 3)
newPoly <- geo_translate(obj = gtGeoms$polygon, x = 5, fid = 2,
                        update = FALSE)
geo_vis(geom = newPoly, linecol = "green", new = FALSE)
}
\seealso{
Other geometry tools: 
\code{\link{geo_filter}()},
\code{\link{geo_locate}()},
\code{\link{geo_pull}()},
\code{\link{geo_reflect}()},
\code{\link{geo_rotate}()},
\code{\link{geo_scale}()},
\code{\link{geo_skew}()},
\code{\link{geo_stretch}()}
}
\concept{geometry tools}
