% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_vis.R
\name{geo_vis}
\alias{geo_vis}
\title{Visualise geometric objects}
\usage{
geo_vis(
  ...,
  window = NULL,
  trace = FALSE,
  new = TRUE,
  clip = TRUE,
  theme = geoTheme
)
}
\arguments{
\item{...}{objects to plot and optional graphical parameters.}

\item{window}{\link[=data.frame]{data.frame(2)}\cr two opposing corners of a rectangle
to which the plot is limited.}

\item{trace}{\link[=logical]{logical(1)}\cr Print the provenance information of the
geometric object (if available) (\code{TRUE}), or simply plot the object
(\code{FALSE}, default).}

\item{new}{\link[=logical]{logical(1)}\cr force a new plot (\code{TRUE}, default).}

\item{clip}{\link[=logical]{logical(1)}\cr clip the plot by the plot box
(\code{TRUE}, default), or plot also objects that go beyond the plot box.}

\item{theme}{\link[=geoTheme]{geometr theme}\cr the theme from which to take graphical
parameters; see \code{\link{geo_theme}} for details.}
}
\value{
Returns invisibly an object of class \code{recordedplot}, see
\code{\link{recordPlot}} for details (and warnings).
}
\description{
Visualise geometric objects
}
\examples{
# make an empty plot
geo_vis()
geo_vis(window = getExtent(gtGeoms$grid))

coords <- data.frame(x = c(30, 60, 60, 40),
                     y = c(40, 40, 60, 70),
                     fid = 1)
(aGeom <- gs_polygon(anchor = coords))
geo_vis(aGeom)

win <- data.frame(x = c(0, 80),
                  y = c(0, 80))
withWindow <- geomio::setWindow(x = aGeom, to = win)
geo_vis(expanded = withWindow)

(aRaster <- gtGeoms$grid)

# plot several objects together ...
geo_vis(aRaster, aGeom)

# ... and give them names
geo_vis(`a raster` = aRaster, `a geom` = aGeom)

# use graphical parameters ...
geo_vis(aGeom, linecol = "green")

# ... or a theme
geo_vis(aRaster, theme = geo_theme(title = list(plot = FALSE)))

}
