% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1geom.R
\docType{class}
\name{geom-class}
\alias{geom-class}
\alias{geom}
\title{Geometry class (S4) and methods}
\arguments{
\item{window}{\link[=data.frame]{data.frame(2)}\cr in case the reference window
deviates from the bounding box of \code{crds}, specify here the minimum and
maximum values in columns \code{x} and \code{y}.}
}
\description{
A \code{geom} stores a table of points, a table of feature to which the
points are associated and a table of groups, to which features are
associated. A \code{geom} can be spatial (if it has a coordinate reference
system assigned to it), but is not by default.
}
\details{
A \code{geom} has one of three geometry types: \itemize{ \item
\code{"point"}, when none of the points are connected to other points, \item
\code{"line"}, where points with the same \code{fid} are connected following
the sequence of their order, without the line closing in itself and \item
\code{"polygon"}, where points with the same \code{fid} are connected
following the sequence of their order and the line closes in on itself due to
first and last point being the same. Moreover, \code{polygon} objects can
contain holes.}

The data model for storing points follows the spaghetti model. Points are
stored as a sequence of x and y values, associated to a feature ID. The
feature ID relates coordinates to features and thus common attributes. Points
and Lines are implemented straightforward in this model, but polygons, which
may contain holes, are a bit trickier. In \code{geometr} they are implemented
as follows: \enumerate{ \item All points with the same \code{fid} make up one
polygon, irrespective of it containing holes or not. \item The outer
path/ring of a polygon is composed of all points until a duplicated of its
first point occurs. This signals that all following points are part of
another path/ring, which must be inside the outer path and which consists of
all points until a duplicate of it's first point occurs. \item This repeats
until all points of the feature are processed.}

Moreover, a \code{geom} has a \emph{reference window}, which is sort of a
second extent that may be bigger (or smaller) than the extent and which
determines the relative position of the points when plotting.
}
\section{Slots}{

\describe{
\item{\code{type}}{\link[=character]{character(1)}\cr the type of feature, either
\code{"point"}, \code{"line"}, \code{"polygon"} or \code{"grid"}.}

\item{\code{geometry}}{\link[=data.frame]{data.frame(3)}\cr the \code{fid} (feature ID),
\code{x} and \code{y} coordinates per point and optional arbitrary point
attributes.}

\item{\code{data}}{\link[=list]{named list}\cr A list with the layer name and list elements
\code{$features} and \code{$groups} containing the features and groups of
that layer. \itemize{
\item features: \code{fid} (feature ID), \code{gid} (group ID) and optional
arbitrary feature attributes.
\item groups: \code{gid} (group ID) and optional arbitrary group attributes.
}}

\item{\code{crs}}{\link[=character]{character(1)}\cr the coordinate reference system in
proj4 notation.}

\item{\code{provenance}}{\link[=list]{list(.)}\cr a list of steps taken to derive the
\code{geom} in focus.}
}}

