% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet-hydrometric.R
\name{geomet_annual_stats}
\alias{geomet_annual_stats}
\title{Query Water Survey of Canada data}
\usage{
geomet_annual_stats(station_number = NULL, as_spatial = FALSE)
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number.}

\item{as_spatial}{Convert to sf object. Defaults to FALSE.}
}
\description{
Query geomet webservice for hydrometric data
}
\examples{
 geomet_annual_stats("08MF005")
}
