% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet.R
\name{geomet_collection}
\alias{geomet_collection}
\title{Collections available in the GeoMet webservice}
\usage{
geomet_collection()
}
\description{
Outputs a tibble of the collections currently available via the webservice.
}
