% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet.R
\name{geomet_data}
\alias{geomet_data}
\title{Query GeoMet for data}
\usage{
geomet_data(parameter, station_number = NULL, as_spatial = TRUE,
  start_date = NULL, end_date = NULL, page_limit = 500,
  verbose = FALSE)
}
\arguments{
\item{parameter}{One of \code{geomet_collection()$name}}

\item{station_number}{A meteorological/hydrometric station number.}

\item{as_spatial}{Convert to sf object. Defaults to TRUE.}

\item{start_date}{Accepts either YYYY-MM-DD. If this argument is left missing, the function will return all data prior to \code{end_date}. If both are NULL all
available data is returned.}

\item{end_date}{Accepts either YYYY-MM-DD. If this argument is left missing, the function will return all data after \code{start_date}. If both are NULL all
available data is returned.}

\item{page_limit}{Default to 500. You likely will not need to change this value. This represents the maximum number of records that can be requested
for each pagination request.}

\item{verbose}{Defaults to FALSE. Do you want the URLs used to make the GET call to be printed?}
}
\description{
Query geomet webservice for ECCC data
}
\examples{
geomet_data(parameter = "hydrometric-daily-mean", station_number = "10PC003", start_date = "1983-07-04")

## Get all ahccd-stations
geomet_data("ahccd-stations")

geomet_data("ahccd-monthly", station_number = "1171020")
}
