% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometr-package.R
\docType{package}
\name{geometr-package}
\alias{geometr}
\alias{geometr-package}
\title{geometr: ECCC Met data in R}
\description{
An R interface to Environment and Climate Change Canada's GeoMet Webservice.
}
\author{
\strong{Maintainer}: Sam Albers \email{Sam.Albers@gov.bc.ca}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
