% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoms}
\alias{geoms}
\title{Example \code{geom} objects}
\format{
The list contains five objects of class geom, a \code{point},
\code{line} and \code{polygon} object and two \code{grid} geoms (that have
56 by 60 cells), one with categorical values (land-use classes) and one
with continuous values (vegetation cover). They are mostly used in the
example and test-sections of this package.
}
\usage{
geoms
}
\description{
A set of five geometries.
}
\keyword{datasets}
