% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoTheme.R
\docType{class}
\name{geoTheme-class}
\alias{geoTheme-class}
\alias{themeClass}
\title{Theme class (S4) and methods}
\description{
An \code{geoTheme} stores a theme to \code{\link{visualise}} vector and gridded
objects. It is recommended to use \code{\link{setTheme}} to modify a
\code{geoTheme}, because it carries out all the checks and makes sure that
names of the parameters are properly matched.
}
\section{Slots}{

\describe{
\item{\code{title}}{\link[=list]{named list(3)}\cr properties of the title.}

\item{\code{box}}{\link[=list]{named list(5)}\cr properties of the bounding box.}

\item{\code{xAxis}}{\link[=list]{named list(5)}\cr properties of the x-axis, its labels
and ticks.}

\item{\code{yAxis}}{\link[=list]{named list(5)}\cr properties of the y-axis, its labels
and ticks.}

\item{\code{grid}}{\link[=list]{named list(5)}\cr properties of the major and minor grid.}

\item{\code{legend}}{\link[=list]{named list(9)}\cr properties of the legend, its title,
labels, ticks and bounding box.}

\item{\code{scale}}{\link[=list]{named list(6)}\cr properties of scaling parameters to
attributes/variables.}

\item{\code{parameters}}{\link[=list]{named list(7)}\cr parameters of the plot object.}
}}

