% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoTheme}
\alias{geoTheme}
\title{Default visualising theme}
\format{
An object of class \code{geoTheme} of length 1.
}
\usage{
geoTheme
}
\description{
Default visualising theme
}
\keyword{datasets}
