% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtent.R
\docType{methods}
\name{getExtent}
\alias{getExtent}
\alias{getExtent,geom-method}
\alias{getExtent,Spatial-method}
\alias{getExtent,sf-method}
\title{Get the extent}
\usage{
\S4method{getExtent}{geom}(x)

\S4method{getExtent}{Spatial}(x)

\S4method{getExtent}{sf}(x)
}
\description{
Get the bounding box of a spatial object.
}
