% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistory.R
\docType{methods}
\name{getHistory}
\alias{getHistory}
\alias{getHistory,geom-method}
\alias{getHistory,RasterLayer-method}
\alias{getHistory,RasterBrick-method}
\alias{getHistory,RasterStack-method}
\title{Get the history of a spatial object.}
\usage{
\S4method{getHistory}{geom}(x)

\S4method{getHistory}{RasterLayer}(x)

\S4method{getHistory}{RasterBrick}(x)

\S4method{getHistory}{RasterStack}(x)
}
\arguments{
\item{x}{the object from which to derive the history.}
}
\description{
Get the history of a spatial object.
}
