% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\alias{getLayer,ANY-method}
\alias{getLayer,geom-method}
\alias{getLayer,Spatial-method}
\alias{getLayer,sf-method}
\alias{getLayer,Raster-method}
\alias{getLayer,matrix-method}
\title{Get a specific layer of a spatial object.}
\usage{
\S4method{getLayer}{ANY}(x)

\S4method{getLayer}{geom}(x, layer = NULL)

\S4method{getLayer}{Spatial}(x, layer = NULL)

\S4method{getLayer}{sf}(x, layer = NULL)

\S4method{getLayer}{Raster}(x, layer = NULL)

\S4method{getLayer}{matrix}(x, layer = NULL)
}
\arguments{
\item{x}{the object from which to get the layer.}

\item{layer}{[\code{character(.)} | \code{integerish(.)}]\cr the layer(s) to get.}
}
\value{
A list of the requested layers.
}
\description{
Get a specific layer of a spatial object.
}
\seealso{
Other getters: 
\code{\link{getCRS}()},
\code{\link{getExtent}()},
\code{\link{getFeatures}()},
\code{\link{getGroups}()},
\code{\link{getHistory}()},
\code{\link{getName}()},
\code{\link{getPoints}()},
\code{\link{getRes}()},
\code{\link{getType}()},
\code{\link{getWindow}()}
}
\concept{getters}
