% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getName.R
\name{getName}
\alias{getName}
\alias{getName,ANY-method}
\alias{getName,geom-method}
\alias{getName,sf-method}
\alias{getName,Raster-method}
\title{Get the name(s) of a spatial object.}
\usage{
\S4method{getName}{ANY}(x)

\S4method{getName}{geom}(x)

\S4method{getName}{sf}(x)

\S4method{getName}{Raster}(x)
}
\arguments{
\item{x}{the object from which to get the name.}
}
\value{
A vector of the requested names.
}
\description{
Get the name(s) of a spatial object.
}
\seealso{
Other getters: 
\code{\link{getCRS}()},
\code{\link{getExtent}()},
\code{\link{getFeatures}()},
\code{\link{getGroups}()},
\code{\link{getHistory}()},
\code{\link{getLayer}()},
\code{\link{getPoints}()},
\code{\link{getRes}()},
\code{\link{getType}()},
\code{\link{getWindow}()}
}
\concept{getters}
