% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getType.R
\docType{methods}
\name{getType}
\alias{getType}
\alias{getType,geom-method}
\alias{getType,Spatial-method}
\alias{getType,sf-method}
\alias{getType,RasterLayer-method}
\title{Get the type of an object.}
\usage{
\S4method{getType}{geom}(x)

\S4method{getType}{Spatial}(x)

\S4method{getType}{sf}(x)

\S4method{getType}{RasterLayer}(x)
}
\arguments{
\item{x}{the object for which to determine the type.}
}
\description{
Get the type of an object.
}
\examples{
getType(gtGeoms$polygon)

getType(x = gtSP$SpatialPolygons)

getType(gtSF$multiline)

getType(gtRasters$categorical)
}
