% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVertices.R
\docType{methods}
\name{getVertices}
\alias{getVertices}
\alias{getVertices,geom-method}
\alias{getVertices,Spatial-method}
\alias{getVertices,sf-method}
\alias{getVertices,ppp-method}
\title{Get the table of coordinates of a spatial object.}
\usage{
\S4method{getVertices}{geom}(x)

\S4method{getVertices}{Spatial}(x)

\S4method{getVertices}{sf}(x)

\S4method{getVertices}{ppp}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinates}
}
\value{
A table of the coordinates \code{x} is made up of.
}
\description{
Get the table of coordinates of a spatial object.
}
\examples{
getVertices(gtGeoms$polygon)

getVertices(gtSP$SpatialPoints)

getVertices(gtSF$multilinestring)

getVertices(gtPPP)
}
\seealso{
Other getters: \code{\link{getCRS}},
  \code{\link{getExtent}}, \code{\link{getHistory}},
  \code{\link{getSubset}}, \code{\link{getTable}},
  \code{\link{getType}}, \code{\link{getWindow}}
}
\concept{getters}
