% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_point.R
\name{gs_point}
\alias{gs_point}
\title{Create a point geometry}
\usage{
gs_point(anchor = NULL, window = NULL, template = NULL,
  vertices = NULL, show = FALSE, ...)
}
\arguments{
\item{anchor}{[\code{data.frame(1)}]\cr Object to derive the \code{geom}
from. It must include column names \code{x}, \code{y} and optinal variables
such as \code{fid}; see Examples.}

\item{window}{[\code{data.frame(1)}]\cr in case the reference window deviates
from the bounding box of \code{anchor} (minimum and maximum values),
specify this here.}

\item{template}{[\code{RasterLayer(1)} | \code{matrix(1)}]\cr Gridded object
that serves as template to sketch the geometry.}

\item{vertices}{[\code{integer(1)}]\cr number of vertices.}

\item{show}{[\code{logical(1)}]\cr in case \code{template} is set, should the
points be plotted (\code{TRUE}) or should it not be plotted (\code{FALSE},
default)?}

\item{...}{[various]\cr graphical parameter, in case \code{show = TRUE}; see
\code{\link{gpar}}.}
}
\value{
An invisible \code{geom}.
}
\description{
Create a point geometry
}
\examples{
# create points programmatically
somePoints <- data.frame(X = c(5190599, 5222810, 5041066, 5234735,
                         5326537, 5027609, 5281527, 5189955), Y = c(3977612,
                         4060164, 3997230, 4117856, 4028167, 3971119, 4118207,
                         4062838))
(pointsGeom <- gs_point(anchor = somePoints, col = "darkorange", show = TRUE))

\dontrun{

input <- rtRasters$continuous

# create points interactively
myPoints <- gs_point(template = input, vertices = 5, show = TRUE, col = "deeppink")
myPoints <- geom_group(myPoints, index = rep(1, 5))
anExtent <- gs_ractangle(myPoints, show = TRUE, col = "green")
}
}
\seealso{
Other shapes: \code{\link{gs_polygon}},
  \code{\link{gs_random}}
}
\concept{shapes}
