% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_random.R
\name{gs_random}
\alias{gs_random}
\title{Create a geometry randomly}
\usage{
gs_random(type = "point", template = NULL, vertices = NULL,
  show = FALSE, ...)
}
\arguments{
\item{type}{[\code{character(1)}]\cr Either one of the three main feature
types \code{"point"}, \code{"line"} or \code{"polygon"}, or more
specifically one of their subtypes, e.g. \code{"hexagon"}.}

\item{template}{[\code{RasterLayer(1)} | \code{matrix(1)}]\cr Gridded object
that serves as template to sketch the geometry.}

\item{vertices}{[\code{integerish(1)}]\cr the number of vertices the geometry
should have; only meaningful if \code{type} does not indicate the number of
vertices already. If left at \code{NULL} the minimum number of vertices for
the \code{geom} type, i.e. 1 for \code{point}, 2 for \code{line} and 3 for
\code{polygon}}

\item{show}{[\code{logical(1)}]\cr should the geometry be plotted
(\code{TRUE}) or should it not be plotted (\code{FALSE}, default)? In case
\code{template} is set, it is automatically \code{TRUE}.}

\item{...}{[various]\cr graphical parameter, in case \code{show = TRUE}; see
\code{\link{gpar}}.}
}
\description{
This function creates a random geometry
}
\examples{
input <- matrix(nrow = 100, ncol = 100, data = 0)

# create a random geometry with five vertices
set.seed(1)
someGeom <- geomRand(type = "polygon", vertices = 5)
visualise(geom = someGeom)

# in case template is given, this serves as source for the window extent
someGeom <- geomRand(template = input, col = "green", show = TRUE)
}
\seealso{
Other shapes: \code{\link{gs_point}},
  \code{\link{gs_polygon}}
}
\concept{shapes}
