% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_geom.R
\name{gt_as_geom}
\alias{gt_as_geom}
\title{Derive a \code{geom} from other spatial objects}
\usage{
gt_as_geom(input)
}
\arguments{
\item{input}{[\code{Spatial*} | \code{sf}]\cr the spatial object to build a
\code{geom} from.}
}
\value{
a \code{geom} of the type that comes closest to the type of the
  input.
}
\description{
See \code{\link{geom}} for details on the class.
}
\details{
bla
}
\examples{
library(raster)

spPoly <- gtSP$SpatialPolygons
plot(spPoly)
visualise(geom = gt_as_geom(input = spPoly))

sfPoly <- gtSF$polygon
plot(sfPoly)
visualise(geom = gt_as_geom(input = sfPoly))

}
