% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_grob.R
\name{gt_as_grob}
\alias{gt_as_grob}
\title{Transform a \code{geom} to a grob object}
\usage{
gt_as_grob(geom = NULL, theme = gtTheme, ...)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{theme}{[\code{list(7)}]\cr Visualising options; see}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  \code{\link{polylineGrob}} or a \code{\link{pathGrob}}.
}
\description{
An object of class \code{\link{grob}} is the grid-package representation of a
\code{geom} and is used for plotting.
}
