% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_raster.R
\name{gt_as_raster}
\alias{gt_as_raster}
\title{Transform geometry to raster}
\usage{
gt_as_raster(geom, negative = FALSE, res = c(1, 1), crs = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{negative}{[\code{logical(1)}]\cr should the area covered by \code{geom}
be set to 0 (\code{TRUE}) or should it be set to 1 (\code{FALSE}, default)?}

\item{res}{[\code{numeric(2)}]\cr resolution in x and y direction.}

\item{crs}{[\code{character(1)}]\cr corrdinate reference system of the
object in proj4 notation.}
}
\value{
a binary \code{\link{raster}} with the dimensions of the reference
  window of \code{geom} and the resolution \code{res}.
}
\description{
An object of class \code{RasterLayer} is the raster-package representation of
a \code{geom}.
}
\examples{
coords <- data.frame(x = c(40, 70, 70, 50),
                     y = c(40, 40, 60, 70),
                     fid = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window)

aRaster <- gToRaster(geom = aGeom)
visualise(raster = aRaster, geom = aGeom, col = "deeppink")

negRaster <- gToRaster(geom = aGeom, negative = TRUE)
visualise(raster = negRaster, geom = aGeom, col = "deeppink")
}
