% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_sf.R
\name{gt_as_sf}
\alias{gt_as_sf}
\title{Transform a \code{geom} to an sf object}
\usage{
gt_as_sf(geom = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}
}
\value{
If \code{geom} has attributes other than \code{fid} and \code{gid} a
  "Simple feature collection", otherwise a "Geometry set". Several features
  of the \code{geom} are returned as MULTI* feature, when they have
  \code{gid} and optionally other attributes in common, otherwise they are
  returned as a single simple feature.
}
\description{
An object of class \code{sf} is the sf-package representation of a
\code{geom}.
}
\examples{
(sfPoints <- gt_as_sf(geom = gtGeoms$point))
(sfLines <- gt_as_sf(geom = gtGeoms$line))
(sfPolygon <- gt_as_sf(geom = gtGeoms$polygon))
}
