% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_edit.R
\name{gt_edit}
\alias{gt_edit}
\title{Edit point geometries}
\usage{
gt_edit(panel = NULL, tolerance = 0.01, verbose = FALSE,
  centroid = FALSE)
}
\arguments{
\item{panel}{[\code{character(1)}]\cr the panel in which to edit vertices
(i.e. the title shown over the plot). Does not have to be accurate, as it
matches with \code{\link{grep}}.}

\item{tolerance}{[\code{numeric(1)}]\cr width of the locator-boxes as}

\item{verbose}{[\code{logical(1)}]\cr be verbose about intermediate steps?}

\item{centroid}{[\code{logical(1)}]\cr instead of moving the vertices of a
\code{geom}, move the centroid and thus the overall \code{geom}. proportion
of the overall box width.}
}
\description{
\code{gt_edit} allows to edit the vertices of a \code{geom} of type
\code{point}.
}
\examples{
\dontrun{
library(sf)

nc = st_read(system.file("shape/nc.shp", package="sf"))
nc_geom <- gt_sf(input = nc)
visualise(geom = nc_geom, window = data.frame(x = c(-82.1, -81), y = c(36.6, 35.9)))
modGeom <- gt_edit()
}
}
