% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_grob.R
\name{gt_grob}
\alias{gt_grob}
\title{Transform a \code{geom} to and from a \code{grob} object}
\usage{
gt_grob(input = NULL, window = NULL, theme = gtTheme, ...)
}
\arguments{
\item{input}{[\code{geom}]\cr Object to transform.}

\item{window}{[\code{data.frame(1)}]\cr in case the reference window deviates
from the bounding box of \code{input} (minimum and maximum values),
specify this here.}

\item{theme}{[\code{list(7)}]\cr visualising options; see}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  \code{\link{polylineGrob}} or a \code{\link{pathGrob}}.
}
\description{
An object of class \code{\link{grob}} is the grid-package representation of a
\code{geom} and is used for plotting.
}
\seealso{
Other spatial classes: \code{\link{gt_sf}},
  \code{\link{gt_sp}}
}
\concept{spatial classes}
