% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_group.R
\name{gt_group}
\alias{gt_group}
\title{Group the vertices of geometries}
\usage{
gt_group(geom, index = NULL, distance = NULL, clusters = NULL, ...)
}
\arguments{
\item{geom}{[\code{geom}]\cr object of class \code{\link{geom}} and type
\code{point}.}

\item{index}{[\code{integerish(.)}]\cr a vector with group values for all
vertices.}

\item{distance}{[\code{numeric(1)}]\cr specific distance of two vertices
below which they will be included in the same cluster.}

\item{clusters}{[\code{integerish(1)}]\cr the number of clusters for
\code{\link{kmeans}} clustering.}

\item{...}{[various]\cr additional arguments either to
\code{\link[stats]{hclust}} or to \code{\link[stats]{kmeans}}.}
}
\value{
\code{geom} with (re)grouped coordinates.
}
\description{
\code{gt_group} assigns the vertices of a \code{geom} into new groups.
}
\details{
Only one of the three arguments \code{index}, \code{distance} or
  \code{clusters} needs to be set, as grouping is only carried out by one of
  them.
}
\examples{
coords <- data.frame(x = c(30, 60, 60, 40, 10, 40, 20),
                     y = c(40, 40, 60, 70, 10, 20, 40),
                     fid = c(1, 1, 2, 2, 2, 2, 2))
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- gs_point(anchor = coords, window = window)

grouped <- gt_group(geom = aGeom, distance = 40)
visualise(geom = grouped)
}
\seealso{
Other geometry tools: \code{\link{gt_edit}},
  \code{\link{gt_rotate}}, \code{\link{gt_scale}},
  \code{\link{gt_skew}}, \code{\link{gt_stretch}}
}
\concept{geometry tools}
