% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sf.R
\name{gt_sf}
\alias{gt_sf}
\title{Transform a \code{geom} to and from an \code{sf} object}
\usage{
gt_sf(input = NULL)
}
\arguments{
\item{input}{[\code{geom} | \code{sf}]\cr Object to transform.}
}
\value{
If \code{input} is a \code{geom} and has attributes other than
  \code{fid} and \code{gid}, a "Simple feature collection", otherwise a
  "Geometry set". Several features of the \code{geom} are returned as MULTI*
  feature, when they have \code{gid} and optionally other attributes in
  common, otherwise they are returned as a single simple feature.

  If \code{input} is a \code{sf}, a \code{geom} that reflects the simple
  feature type.
}
\description{
An object of class \code{sf} is the sf-package representation of a
\code{geom}.
}
\examples{
(sfPoints <- gt_sf(input = gtGeoms$point))
(sfLines <- gt_sf(input = gtGeoms$line))
(sfPolygon <- gt_sf(input = gtGeoms$polygon))

sfPoly <- gtSF$polygon
plot(sfPoly)
visualise(geom = gt_sf(input = sfPoly))
}
\seealso{
Other spatial classes: \code{\link{gt_grob}},
  \code{\link{gt_sp}}
}
\concept{spatial classes}
