% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sp.R
\name{gt_sp}
\alias{gt_sp}
\title{Transform a \code{geom} to and from a \code{Spatial*} object}
\usage{
gt_sp(input = NULL)
}
\arguments{
\item{input}{[\code{geom} | \code{sp}]\cr Object to transform.}
}
\value{
Depending on \code{input} either a \code{geom} or a \code{sp}
  reflecting the type of \code{input}.
}
\description{
An object of class \code{Spatial*} is the sp-package representation of a
\code{geom}.
}
\examples{
(spPoints <- gt_sp(input = gtGeoms$point))
(spLines <- gt_sp(input = gtGeoms$line))
(spPolygon <- gt_sp(input = gtGeoms$polygon))

spPoly <- gtSP$SpatialPolygons
sp::plot(spPoly)
visualise(geom = gt_sp(input = spPoly))
}
\seealso{
Other spatial classes: \code{\link{gt_grob}},
  \code{\link{gt_raster}}, \code{\link{gt_sf}}
}
\concept{spatial classes}
