% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_vEdit.R
\name{gt_vEdit}
\alias{gt_vEdit}
\title{Edit point geometries}
\usage{
gt_vEdit(geom)
}
\arguments{
\item{geom}{[\code{geom}]\cr object of class \code{\link{geom}} and type
\code{point}.}
}
\description{
\code{gt_vEdit} allows to edit the vertices of a \code{geom} of type
\code{point}.
}
