% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtPPP}
\alias{gtPPP}
\title{Example \code{ppp} object}
\format{
The object contains an object of class \code{ppp}. It consists of 15
  points without an attribute table. It is mostly used in the example and
  test-sections of this package.
}
\usage{
gtPPP
}
\description{
A single ppp object
}
\keyword{datasets}
