% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtRasters}
\alias{gtRasters}
\title{Example \code{RasterStack} object}
\format{
The object of class RasterStack has no projection and is a
  RasterStack object of 56 by 60 cells. The first raster represents
  land-use classes and the second raster contains a continuous scale of
  vegetation cover.
}
\usage{
gtRasters
}
\description{
A set of two conceptually different types of raster.
}
\keyword{datasets}
