% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtSF}
\alias{gtSF}
\title{Example \code{sf} objects}
\format{
The list contains six objects of class \code{sf}, a \code{POINT}, a
  \code{MULTIPOINT}, a \code{LINESTRING}, a \code{MULTILINESTRING}, a
  \code{POLYGON}, and a \code{MULTIPOLYGON} object. They are mostly used in
  the example and test-sections of this package.
}
\usage{
gtSF
}
\description{
A set of six sp objects
}
\keyword{datasets}
