% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtSP}
\alias{gtSP}
\title{Example \code{Spatial} objects}
\format{
The list contains four objects of class \code{Spatial}, a
  \code{SpatialPoints}, a \code{SpatialMultiPoints}, a, \code{SpatialLines}
  and a \code{SpatialPolygons} object. They are mostly used in the example
  and test-sections of this package.
}
\usage{
gtSP
}
\description{
A set of four sp objects
}
\keyword{datasets}
