% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtTheme}
\alias{gtTheme}
\title{Default visualising theme}
\format{
An object of class \code{gtTheme} of length 1.
}
\usage{
gtTheme
}
\description{
Default visualising theme
}
\keyword{datasets}
