% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{methods}
\name{length,geom-method}
\alias{length,geom-method}
\title{Determin number of vertices}
\usage{
\S4method{length}{geom}(x)
}
\arguments{
\item{x}{[\code{geom}]\cr object from which to determine \code{length}.}
}
\description{
Determin number of vertices
}
