% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listArgs}
\alias{listArgs}
\title{List arguments of the parent function}
\usage{
listArgs()
}
\value{
The output of \code{\link{match.call}} for the parent frame.
}
\description{
Determine the name and value of the arguments called in a function.
}
\examples{

testFun <- function(x = "character", y = NULL, z = c(1, 2, 3)){
  g <- listArgs()
  return(g)
}
testFun(x = "hello world")
}
