% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{load_csv}
\alias{load_csv}
\title{Load \code{csv} files}
\usage{
load_csv(path)
}
\arguments{
\item{path}{[\code{character(1)}]\cr the local path of the file to load.}
}
\value{
a \code{point geometry} of the coordinates
}
\description{
csv is the short form of \emph{Comma Separated Values}, which represents
simple tables of systematically delimited information. The files loaded with
this function should have the columns \code{x}, \code{y} and potentially
other columns that represent the attributes of these coordinates, for example
an \code{id}.
}
\details{
A column \code{fid} can have a deviating value than \code{id} when several
features are part of the same object, such as multiple line features or holes
in polygons.
}
\seealso{
Other loaders: \code{\link{load_kml}}
}
\concept{loaders}
