% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Locate (and identify) clicks}
\usage{
locate(samples = 1, raster = NULL, panel = NULL, identify = FALSE,
  snap = FALSE, raw = FALSE, silent = FALSE, show = FALSE, ...)
}
\arguments{
\item{samples}{[\code{integerish(1)}]\cr the number of clicks.}

\item{raster}{[\code{Raster*} | \code{matrix}]\cr the raster object in which
to locate.}

\item{panel}{[\code{character(1)}]\cr the panel in which to locate (i.e. the
title shown over the plot). Does not have to be accurate, as it matches
with \code{\link{grep}}.}

\item{identify}{[\code{logical(1)}]\cr get the raster value at the sampled
location (\code{TRUE}) or merely the location (\code{FALSE}, default).}

\item{snap}{[\code{logical(1)}]\cr should the returned value(s) be set to the
nearest grid cell's center (\code{TRUE}) or should they remain the
selected, "real" value (\code{FALSE}, default)?}

\item{raw}{[\code{logical(1)}]\cr should the complete statistics about the
clicks be returned (\code{TRUE}), or should only the basic output be
returned (\code{FALSE}, default)?}

\item{silent}{[\code{logical(1)}]\cr should all messages except errors be
suppressed (\code{TRUE}), or should all messages be printed (\code{FALSE},
default)?}

\item{show}{[\code{logical(1)}]\cr should information about the selected
cells be included in the plot (\code{TRUE}), or should they merely be
returned in the console (\code{FALSE, default})?}

\item{...}{[\code{various}]\cr graphical parameters of the objects that are
created when \code{show = TRUE}.}
}
\value{
a \code{data.frame} of the selected locations and, if \code{identify
  = TRUE}, the respective values. If \code{show = TRUE} the values are also
  shown in the plot.
}
\description{
Click into a plot to get the location or identify values
}
\examples{
\dontrun{
input <- rtRasters$continuous
patches <- rPatches(rBinarise(input, thresh = 30))

# locate from a not yet plotted object
locate(samples = 3, raster = input, show = TRUE, col = "green")

# or from one that is already plotted
visualise(raster = input)
locate(identify = TRUE, snap = TRUE)

# with several panels, specify a target
visualise(raster = raster::brick(input, patches))
locate(samples = 4, panel = "patches", identify = TRUE, show = TRUE)
}
}
