% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeFormat}
\alias{makeFormat}
\title{Gather the format of the plot}
\usage{
makeFormat(panelExt = NULL, theme = NULL)
}
\arguments{
\item{panelExt}{[]}

\item{theme}{[]}
}
\description{
Gather the format of the plot
}
