% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeLayout}
\alias{makeLayout}
\title{Gather the format of the plot}
\usage{
makeLayout(x = NULL, window = NULL, theme = NULL, ...)
}
\arguments{
\item{x}{[]}

\item{window}{[]}

\item{theme}{[]}

\item{...}{[]}
}
\description{
Gather the format of the plot
}
