% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeObject.R
\docType{methods}
\name{makeObject}
\alias{makeObject}
\alias{makeObject,geom-method}
\alias{makeObject,Raster-method}
\alias{makeObject,matrix-method}
\alias{makeObject,Spatial-method}
\alias{makeObject,sf-method}
\title{Make the object to a plot}
\usage{
\S4method{makeObject}{geom}(x, theme, ...)

\S4method{makeObject}{Raster}(x, theme, image = FALSE, ...)

\S4method{makeObject}{matrix}(x, theme, image = FALSE, ...)

\S4method{makeObject}{Spatial}(x, theme, ...)

\S4method{makeObject}{sf}(x, theme, ...)
}
\arguments{
\item{x}{the object from which to make object to plot.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}

\item{image}{[\code{logical(1)}]\cr whether or not the raster (stack)
contains an image.}
}
\description{
Make the object to a plot
}
