% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projs}
\alias{projs}
\title{Projections which are in use in lomm}
\format{List with 5 elemens. \describe{ \item{\code{laea}}{Lambert azimuthal
  equal-area projection,\cr Based on the Geodetic Reference System 1980 (GRS
  80) Ellipsoid} \item{\code{longlat}}{Longitude/Latitude "projection",\cr
  Based on the World Geodetic System 1985 (WGS 84) Ellipsoid}
  \item{\code{utm}}{Universal Transverse Mercator "projection", zone 32,\cr
  Based on the World Geodetic System 1985 (WGS 84) Ellipsoid}
  \item{\code{sinu}}{Sinusoidal projection,\cr Based on a pseudocylinder}
  \item{\code{tmerc}}{Transverse Mercator projection,\cr Based on the Bessel
  ellipsoid} }}
\usage{
projs
}
\description{
Projections which are in use in lomm
}
\keyword{datasets}
