% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rad}
\alias{rad}
\title{Transform degree to radians}
\usage{
rad(degree)
}
\arguments{
\item{degree}{[\code{numeric(1)}]\cr the degree value to transform.}
}
\value{
a radians value
}
\description{
Transform degree to radians
}
