% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCRS.R
\docType{methods}
\name{setCRS}
\alias{setCRS}
\alias{setCRS,geom-method}
\alias{setCRS,Spatial-method}
\alias{setCRS,sf-method}
\title{Set (or transform) the coordinate reference system}
\usage{
\S4method{setCRS}{geom}(x, crs)

\S4method{setCRS}{Spatial}(x, crs)

\S4method{setCRS}{sf}(x, crs)
}
\description{
In case an object has not yet assigned a coordinate reference system, this
function simply assigns it. In case the object has already a valid crs, a
transformation to the new crs will be carried out. The transformation is
computed with the standard defined in the \code{rgdal} package.
}
