% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFeatures.R
\name{setFeatures}
\alias{setFeatures}
\alias{setFeatures,ANY-method}
\alias{setFeatures,geom-method}
\alias{setFeatures,Spatial-method}
\alias{setFeatures,sf-method}
\alias{setFeatures,sfc-method}
\alias{setFeatures,ppp-method}
\title{Set a table of feature attributes}
\usage{
\S4method{setFeatures}{ANY}(x)

\S4method{setFeatures}{geom}(x, table = NULL)

\S4method{setFeatures}{Spatial}(x, table = NULL)

\S4method{setFeatures}{sf}(x, table = NULL)

\S4method{setFeatures}{sfc}(x, table = NULL)

\S4method{setFeatures}{ppp}(x, table = NULL)
}
\arguments{
\item{x}{the object to which to assign a new attribute table.}

\item{table}{[\code{data.frame(.)}]\cr the attribute table.}
}
\value{
The object \code{x} with an updated feature attribute table.
}
\description{
Set a table of feature attributes
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setGroups}()},
\code{\link{setHistory}()},
\code{\link{setWindow}()}
}
\concept{setters}
