% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGroups.R
\docType{methods}
\name{setGroups}
\alias{setGroups}
\alias{setGroups,ANY-method}
\alias{setGroups,geom-method}
\alias{setGroups,RasterLayer-method}
\title{Set the table of group attributes}
\usage{
\S4method{setGroups}{ANY}(x)

\S4method{setGroups}{geom}(x, table = NULL)

\S4method{setGroups}{RasterLayer}(x, table = NULL)
}
\arguments{
\item{x}{the object to which to assign an attribute table.}

\item{table}{[\code{data.frame(.)}]\cr the attribute table.}
}
\value{
The object \code{x} with an updated attribute table.
}
\description{
Set the table of group attributes
}
\seealso{
Other setters: \code{\link{setCRS}},
  \code{\link{setFeatures}}, \code{\link{setHistory}},
  \code{\link{setPoints}}, \code{\link{setWindow}}
}
\concept{setters}
