% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setHistory.R
\docType{methods}
\name{setHistory}
\alias{setHistory}
\alias{setHistory,geom-method}
\alias{setHistory,Raster-method}
\title{Set additional entries to the history of an object}
\usage{
\S4method{setHistory}{geom}(x, history = NULL)

\S4method{setHistory}{Raster}(x, history = NULL)
}
\arguments{
\item{x}{the object for which to set the coordinate reference system.}

\item{history}{[\code{character(1)}]\cr the coordinate reference system to set
for this object.}
}
\value{
bla
}
\description{
Set additional entries to the history of an object
}
