% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\alias{setNames,ANY-method}
\alias{setNames,geom-method}
\title{Set the names of a spatial object.}
\usage{
setNames(x, to, ...)

\S4method{setNames}{ANY}(x)

\S4method{setNames}{geom}(x, to = NULL)
}
\arguments{
\item{x}{the object for which to set a new name.}

\item{to}{new name(s).}

\item{...}{other arguments.}
}
\value{
The object \code{x} with an update name.
}
\description{
Set the names of a spatial object.
}
\examples{
visualise(gtGeoms$grid$categorical)

cover <- setNames(x = gtGeoms$grid$categorical, to = "cover")

visualise(cover)
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setGroups}()},
\code{\link{setHistory}()},
\code{\link{setWindow}()}
}
\concept{setters}
