% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPoints.R
\name{setPoints}
\alias{setPoints}
\alias{setPoints,ANY-method}
\alias{setPoints,geom-method}
\title{Set a table of point attributes}
\usage{
\S4method{setPoints}{ANY}(x, table = NULL)

\S4method{setPoints}{geom}(x, table = NULL)
}
\arguments{
\item{x}{the object to which to assign a new point attribute table.}

\item{table}{[\code{data.frame(.)}]\cr the new attribute table.}
}
\value{
The object \code{x} with an updated point attribute table.
}
\description{
Set a table of point attributes
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setGroups}()},
\code{\link{setHistory}()},
\code{\link{setWindow}()}
}
\concept{setters}
