% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTable.R
\docType{methods}
\name{setTable}
\alias{setTable}
\alias{setTable,geom-method}
\alias{setTable,Spatial-method}
\alias{setTable,sf-method}
\title{Set the attribute table}
\usage{
\S4method{setTable}{geom}(x, table)

\S4method{setTable}{Spatial}(x, table)

\S4method{setTable}{sf}(x, table)
}
\description{
Set the attribute table of a spatial object.
}
