% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWindow.R
\docType{methods}
\name{setWindow}
\alias{setWindow}
\alias{setWindow,geom-method}
\title{Set the reference window of a spatial object.}
\usage{
\S4method{setWindow}{geom}(x, to)
}
\arguments{
\item{x}{the object in which to set the reference window.}

\item{to}{[\code{data.frame(1)}]\cr two oposing corners.}
}
\description{
Set the reference window of a spatial object.
}
\examples{
# create a polygon programmatically
coords <- data.frame(x = c(40, 70, 70, 50),
                     y = c(40, 40, 60, 70))
(aGeom <- gs_polygon(anchor = coords))

window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
(aGeom <- setWindow(x = aGeom, to = window))

visualise(geom = aGeom, new = FALSE)
}
