% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoTheme.R
\name{show,geoTheme-method}
\alias{show,geoTheme-method}
\title{Print geoTheme in the console}
\usage{
\S4method{show}{geoTheme}(object)
}
\arguments{
\item{object}{\link[=geoTheme]{geometr theme}\cr object to \code{show}.}
}
\description{
Print geoTheme in the console
}
