% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtTheme.R
\name{show,gtTheme-method}
\alias{show,gtTheme-method}
\title{Print gtTheme in the console}
\usage{
\S4method{show}{gtTheme}(object)
}
\arguments{
\item{object}{[\code{gtTheme}]\cr object to \code{show}.}
}
\description{
Print gtTheme in the console
}
