% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrob.R
\name{.makeGrob}
\alias{.makeGrob}
\title{Make the grob of a plot}
\usage{
.makeGrob(x, theme = gtTheme)
}
\arguments{
\item{x}{the object to transform to class \code{grob}.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take parameters.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  a \code{\link{polylineGrob}}, a \code{\link{pathGrob}} or a
  \code{\link{rasterGrob}}.
}
\description{
Make the grob of a plot
}
