% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrob.R
\name{.makeGrob}
\alias{.makeGrob}
\title{Make the grob of a plot}
\usage{
.makeGrob(
  x,
  window,
  layout,
  featureType,
  plotValues,
  scaleValues,
  plotParams,
  rows = rows,
  cols = cols,
  theme = gtTheme
)
}
\arguments{
\item{x}{the object to transform to class \code{grob}.}

\item{window}{[\code{data.frame(1)}] two opposing corners of a rectangle to
which the plot is limited.}

\item{layout}{layout to fit the grob into place.}

\item{featureType}{the type of feature to make a grob from.}

\item{plotValues}{the plot values.}

\item{scaleValues}{the scale values.}

\item{plotParams}{new plotting parameters specified
via the quick options in \code{\link{visualise}}.}

\item{rows}{in case it's a grid, the number of rows.}

\item{cols}{in case it's a grid, the number of cols.}

\item{theme}{the theme from which to take parameters.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  a \code{\link{polylineGrob}}, a \code{\link{pathGrob}} or a
  \code{\link{rasterGrob}}.
}
\description{
Make the grob of a plot
}
