% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLayout.R
\name{.makeLayout}
\alias{.makeLayout}
\title{Make the layout of a plot}
\usage{
.makeLayout(x, legend, window, theme = gtTheme)
}
\arguments{
\item{x}{any spatial object to plot.}

\item{legend}{[\code{list(.)}]\cr the legend object built with
\code{\link{.makeLegend}}.}

\item{window}{[\code{data.frame(1)}] two opposing corners of a rectangle to
which the plot is limited.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}
}
\description{
Make the layout of a plot
}
