% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_.R
\name{.makeLayout}
\alias{.makeLayout}
\title{Make the layout of a plot}
\usage{
.makeLayout(x, legend, window, theme = gtTheme)
}
\arguments{
\item{x}{any spatial object to plot.}

\item{legend}{\link[=list]{list(.)}\cr the legend object built with
\code{\link{.makeLegend}}.}

\item{window}{\link[=data.frame]{data.frame(2)}\cr in case the reference window
deviates from the bounding box of \code{crds}, specify here the minimum and
maximum values in columns \code{x} and \code{y}.}

\item{theme}{\link[=gtTheme]{geometr theme}\cr the theme from which to take
graphical parameters.}
}
\description{
Make the layout of a plot
}
