% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{.makeLegend}
\alias{.makeLegend}
\title{Make the legend of a plot}
\usage{
.makeLegend(x, plotParams, theme)
}
\arguments{
\item{x}{[\code{list(1)}]\cr any spatial object to plot.}

\item{plotParams}{[\code{named list(.)}]\cr new plotting parameters specified
via the quick options in \code{\link{visualise}}.}

\item{theme}{[\code{list(7)}]\cr the theme from which to take graphical
parameters.}
}
\description{
Make the legend of a plot
}
