% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{.makeLegend}
\alias{.makeLegend}
\title{Make the legend of a plot}
\usage{
.makeLegend(x, theme)
}
\arguments{
\item{x}{[\code{list(1)}]\cr any spatial object to plot.}

\item{theme}{[\code{character(1)}]\cr the theme from which to take graphical
parameters.}
}
\description{
Make the legend of a plot
}
