% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{.makeLegend}
\alias{.makeLegend}
\title{Make the legend of a plot}
\usage{
.makeLegend(x, scaleValues, plotParams, theme)
}
\arguments{
\item{x}{any spatial object to plot.}

\item{scaleValues}{the scale values.}

\item{plotParams}{new plotting parameters specified
via the quick options in \code{\link{visualise}}.}

\item{theme}{the theme from which to take graphical
parameters.}
}
\description{
Make the legend of a plot
}
