% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{.makeLegend}
\alias{.makeLegend}
\title{Make the legend of a plot}
\usage{
.makeLegend(x = NULL, theme = NULL)
}
\arguments{
\item{x}{[\code{list(.)}]\cr the theme from which to build the legend.}

\item{type}{[\code{character(1)}]\cr whether it is a \code{"raster"} or a
\code{"vector"} legend.}
}
\description{
Make the legend of a plot
}
