% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.rad}
\alias{.rad}
\title{Convert degree to radians}
\usage{
.rad(degree)
}
\arguments{
\item{degree}{[\code{numeric(1)}]\cr a degree value to convert to radians.}
}
\description{
Convert degree to radians
}
