% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{.testPoints}
\alias{.testPoints}
\title{Test points to not contain NA}
\usage{
.testPoints(x, ...)
}
\arguments{
\item{x}{[\code{data.frame}]\cr the points to be tested.}

\item{...}{[\code{.}]\cr additional arguments.}
}
\description{
Test points to not contain NA
}
