% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testTemplate}
\alias{.testTemplate}
\title{Test template for consistency}
\usage{
.testTemplate(x, ...)
}
\arguments{
\item{x}{[\code{RasterLayer | matrix}]\cr the object to be tested for
consistency.}

\item{...}{[\code{.}]\cr additional arguments.}
}
\description{
Test template for consistency
}
