% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.testWindow}
\alias{.testWindow}
\title{Test window for consistency}
\usage{
.testWindow(x, ...)
}
\arguments{
\item{x}{[\code{data.frame}]\cr the object to be tested for
consistency.}

\item{...}{[\code{.}]\cr additional arguments.}
}
\description{
Test window for consistency
}
