% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTheme.R
\name{.updateTheme}
\alias{.updateTheme}
\title{Update the theme according to disply arguments}
\usage{
.updateTheme(x = NULL, newParams = NULL, theme = gtTheme)
}
\arguments{
\item{x}{[\code{list(1)}]\cr any spatial object to plot.}

\item{newParams}{[\code{named list(.)}]\cr new plotting parameters specified
via the quick options in \code{\link{visualise}}.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme that shall be updated.}
}
\description{
Update the theme according to disply arguments
}
