% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateVertices}
\alias{.updateVertices}
\title{Update the vertices}
\usage{
.updateVertices(input = NULL)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr a table of vertices which should be
brought into the correct form.}
}
\description{
Set the vertices in a table so that they are valid for a geom.
}
