% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateWindow}
\alias{.updateWindow}
\title{Update the window}
\usage{
.updateWindow(input = NULL, window = NULL)
}
\arguments{
\item{input}{\link[=data.frame]{data.frame(1)}\cr a table of vertices for which a new
window should be derived.}

\item{window}{\link[=data.frame]{data.frame(2)}\cr in case the reference window
deviates from the bounding box of \code{crds}, specify here the minimum and
maximum values in columns \code{x} and \code{y}.}
}
\value{
A new window that has the extent of input.
}
\description{
Set a window to the minimum/maximum values of input vertices.
}
