% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateWindow}
\alias{.updateWindow}
\title{Update the window}
\usage{
.updateWindow(input = NULL, window = NULL)
}
\arguments{
\item{input}{[\code{data.frame(1)}]\cr a table of vertices for which a new
window should be derived.}

\item{window}{[\code{data.frame(1)}]\cr the old window.}
}
\value{
A new window that has the extent of \code{input}.
}
\description{
Set a window to the minimum/maximum values of input vertices.
}
