% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_geom.R
\docType{methods}
\name{gc_geom}
\alias{gc_geom}
\alias{gc_geom,sf-method}
\alias{gc_geom,Spatial-method}
\title{Make a geometry object of class \code{geom}}
\usage{
\S4method{gc_geom}{sf}(input, window = NULL, ...)

\S4method{gc_geom}{Spatial}(input, window = NULL, ...)
}
\arguments{
\item{input}{the object from which to make an object of class \code{geom}.}

\item{window}{[\code{data.frame(1)}]\cr the reference window of the new geom.}

\item{...}{additional arguments, such as \code{verbose = TRUE/FALSE}.}
}
\value{
an object of class \code{geom}
}
\description{
Make a geometry object of class \code{geom}
}
\examples{
geomPoly <- gc_geom(input = gtSF$polygon)
geomLine <- gc_geom(input = gtSP$SpatialLinesDataFrame)
}
\seealso{
Other spatial classes: \code{\link{gc_grob}},
  \code{\link{gc_sf}}, \code{\link{gc_sp}}
}
\concept{spatial classes}
