% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_grob.R
\name{gc_grob}
\alias{gc_grob}
\alias{gc_grob,geom-method}
\title{Transform a spatial object to a grob}
\usage{
\S4method{gc_grob}{geom}(input = NULL, theme = gtTheme, ...)
}
\arguments{
\item{input}{the object to transform to class \code{grob}.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take parameters.}

\item{...}{instead of providing a modified \code{theme}, you can also
determine specific graphic parameters (see \code{\link{gpar}}) separately;
see \code{\link{setTheme}} for details.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
}
\description{
Transform a spatial object to a grob
}
\seealso{
Other spatial classes: 
\code{\link{gc_geom}()},
\code{\link{gc_ppp}()},
\code{\link{gc_raster}()},
\code{\link{gc_sf}()},
\code{\link{gc_sp}()}
}
\concept{spatial classes}
