% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_sp.R
\name{gc_sp}
\alias{gc_sp}
\alias{gc_sp,geom-method}
\title{Transform a spatial object to class \code{Spatial}}
\usage{
\S4method{gc_sp}{geom}(input = NULL)
}
\arguments{
\item{input}{the object to transform to class \code{Spatial}.}
}
\value{
an object of class \code{Spatial}
}
\description{
Transform a spatial object to class \code{Spatial}
}
\examples{
gc_sp(input = gtGeoms$point)

gc_sp(input = gtGeoms$line)

gc_sp(input = gtGeoms$polygon)
}
\seealso{
Other spatial classes: 
\code{\link{gc_geom}()},
\code{\link{gc_grob}()},
\code{\link{gc_ppp}()},
\code{\link{gc_raster}()},
\code{\link{gc_sf}()}
}
\concept{spatial classes}
