% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{class}
\name{geom-class}
\alias{geom-class}
\alias{geom}
\title{Geometry class (S4) and methods}
\description{
A \code{geom} stores the vertices that outline shape, and all additional
information that characterise a geometry. A \code{geom} can be spatial, but
does not have to be. A \code{geom} can either have absolute or relative
values, where relative values specify the vertex position relative to the
\code{window} slot.
}
\details{
A \code{geom} either has the feature type \itemize{ \item \code{"point"},
when none of the vertices are connected to other vertices, \item
\code{"line"}, when vertices with the same id are connected according to
their order without the line closing in itself and \item \code{"polygon"}
according to the same definition, except that the line closes in itself.}
Moreover, a \code{geom} does NOT have the slot \emph{extent}, which
characterises the minimum and maximum value of the vertex coordinates and
which is thus derived "on the fly" from the vertices. Instead it has a
\emph{reference window}, which is sort of a second extent that may be bigger
than \code{extent} and which determines the relative position of the vertices
when plotting.
}
\section{Slots}{

\describe{
\item{\code{type}}{[\code{character(1)}]\cr the type of feature, recently either
\code{"point"}, \code{"line"} or \code{"polygon"}.}

\item{\code{coords}}{[\code{data.frame(1)}]\cr the \code{fid}, \code{x} and \code{y}
coordinates per vertex.}

\item{\code{attr}}{[\code{data.frame(1)}]\cr arbitrary number of attributes per
unique id in \code{coords}}

\item{\code{window}}{[\code{data.frame(1)}]\cr the minimum and maximum value in x and
y dimension of the reference window in which the \code{geom} dwells.}

\item{\code{scale}}{[\code{character(1)}]\cr whether the vertex coordinates are
stored as \code{"absolute"} values, or \code{"relative"} to \code{window}.}

\item{\code{crs}}{[\code{character(1)}]\cr the coordinate reference system in proj4
notation.}

\item{\code{history}}{[\code{list(.)}]\cr a list of steps taken to derive the
\code{geom} in focus.}
}}

\section{Methods}{
 So far the following methods have been defined: \itemize{
  \item Getters: \cr \code{length}, \code{show}, \code{\link{getCoords}},
  \code{\link{getCRS}}, \code{\link{getExtent}}, \code{\link{getHistory}},
  \code{\link{getWindow}}, \code{\link{getSubset}}, \code{\link{getTable}},
  \item Setters: \cr \code{\link{setCRS}}, \code{\link{setTable}},
  \code{\link{setWindow}} }
}

