% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet-hydrometric.R
\name{geomet_daily_mean}
\alias{geomet_daily_mean}
\title{Query Water Survey of Canada data}
\usage{
geomet_daily_mean(station_number = NULL, as_spatial = TRUE)
}
\arguments{
\item{station_number}{A seven digit Water Survey of Canada station number.}

\item{as_spatial}{Convert to sf object. Defaults to FALSE.}
}
\description{
Query geomet webservice for hydrometric data
}
\examples{
 geomet_daily_mean("08MF005")
}
