% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\docType{methods}
\name{getCoords}
\alias{getCoords}
\alias{getCoords,geom-method}
\alias{getCoords,Spatial-method}
\alias{getCoords,sf-method}
\title{Get the table of coordinates of a spatial object.}
\usage{
\S4method{getCoords}{geom}(x)

\S4method{getCoords}{Spatial}(x)

\S4method{getCoords}{sf}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinates}
}
\description{
Get the table of coordinates of a spatial object.
}
\examples{
# coordinates of ...

# ... a geom
getCoords(gtGeoms$mask)

# ... a Spatial* object
getCoords(gtSP$SpatialPoints)

# ... an sf* object
getCoords(gtSF$multilinestring)
}
