% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRS.R
\docType{methods}
\name{getCRS}
\alias{getCRS}
\alias{getCRS,geom-method}
\alias{getCRS,Spatial-method}
\alias{getCRS,sf-method}
\alias{getCRS,Raster-method}
\title{Get the coordinate reference system of a spatial object.}
\usage{
\S4method{getCRS}{geom}(x)

\S4method{getCRS}{Spatial}(x)

\S4method{getCRS}{sf}(x)

\S4method{getCRS}{Raster}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinate reference system.}
}
\description{
Get the coordinate reference system of a spatial object.
}
