% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRS.R
\name{getCRS}
\alias{getCRS}
\alias{getCRS,ANY-method}
\alias{getCRS,geom-method}
\alias{getCRS,Spatial-method}
\alias{getCRS,sf-method}
\alias{getCRS,ppp-method}
\alias{getCRS,Raster-method}
\title{Get the coordinate reference system of a spatial object.}
\usage{
\S4method{getCRS}{ANY}(x)

\S4method{getCRS}{geom}(x)

\S4method{getCRS}{Spatial}(x)

\S4method{getCRS}{sf}(x)

\S4method{getCRS}{ppp}(x)

\S4method{getCRS}{Raster}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinate reference system.}
}
\value{
The coordinate reference system of \code{x} given as proj4string.
}
\description{
Get the coordinate reference system of a spatial object.
}
\seealso{
Other getters: 
\code{\link{getExtent}()},
\code{\link{getFeatures}()},
\code{\link{getGroups}()},
\code{\link{getHistory}()},
\code{\link{getLayer}()},
\code{\link{getName}()},
\code{\link{getPoints}()},
\code{\link{getRes}()},
\code{\link{getType}()},
\code{\link{getWindow}()}
}
\concept{getters}
