% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtent.R
\name{getExtent}
\alias{getExtent}
\alias{getExtent,ANY-method}
\alias{getExtent,geom-method}
\alias{getExtent,Spatial-method}
\alias{getExtent,sf-method}
\alias{getExtent,Raster-method}
\alias{getExtent,SpatRaster-method}
\alias{getExtent,matrix-method}
\title{Get the extent (bounding box) of a spatial object.}
\usage{
getExtent(x, ...)

\S4method{getExtent}{ANY}(x)

\S4method{getExtent}{geom}(x)

\S4method{getExtent}{Spatial}(x)

\S4method{getExtent}{sf}(x)

\S4method{getExtent}{Raster}(x)

\S4method{getExtent}{SpatRaster}(x)

\S4method{getExtent}{matrix}(x)
}
\arguments{
\item{x}{the object from which to derive the extent.}

\item{...}{other arguments.}
}
\value{
A tibble of the lower left and upper right corner coordinates of the
  extent of \code{x}. This table two columns (x and y) and two rows (minimum
  and maximum).
}
\description{
Get the extent (bounding box) of a spatial object.
}
\examples{

getExtent(gtGeoms$line)

gc_sp(gtGeoms$line) \%>\%
  getExtent()

gc_sf(gtGeoms$line) \%>\%
  getExtent()

gc_raster(gtGeoms$grid$categorical) \%>\%
  getExtent()

gc_terra(gtGeoms$grid$categorical) \%>\%
  getExtent()

getExtent(x = matrix(0, 3, 5))
}
\seealso{
Other getters: 
\code{\link{getCRS}()},
\code{\link{getCols}()},
\code{\link{getFeatures}()},
\code{\link{getGroups}()},
\code{\link{getHistory}()},
\code{\link{getLayers}()},
\code{\link{getNames}()},
\code{\link{getPoints}()},
\code{\link{getRes}()},
\code{\link{getRows}()},
\code{\link{getType}()},
\code{\link{getWindow}()}
}
\concept{getters}
