% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtent.R
\docType{methods}
\name{getExtent}
\alias{getExtent}
\alias{getExtent,geom-method}
\alias{getExtent,Spatial-method}
\alias{getExtent,sf-method}
\alias{getExtent,ppp-method}
\alias{getExtent,Raster-method}
\alias{getExtent,matrix-method}
\title{Get the extent (bounding box) of a spatial object.}
\usage{
\S4method{getExtent}{geom}(x)

\S4method{getExtent}{Spatial}(x)

\S4method{getExtent}{sf}(x)

\S4method{getExtent}{ppp}(x)

\S4method{getExtent}{Raster}(x)

\S4method{getExtent}{matrix}(x)
}
\arguments{
\item{x}{the object from which to derive the extent.}
}
\value{
A table of the lower left and upper right corner of the extent of \code{x}.
}
\description{
Get the extent (bounding box) of a spatial object.
}
\examples{
getTable(gtGeoms$polygon)

getExtent(gtSP$SpatialPolygons)

getExtent(gtSF$multilinestring)

getExtent(gtPPP)

getExtent(gtRasters$categorical)

getExtent(matrix(0, 3, 5))
}
