% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\docType{methods}
\name{getGroups}
\alias{getGroups}
\alias{getGroups,ANY-method}
\alias{getGroups,geom-method}
\alias{getGroups,Raster-method}
\title{Get the table of group attributes}
\usage{
\S4method{getGroups}{ANY}(x, ...)

\S4method{getGroups}{geom}(x, ...)

\S4method{getGroups}{Raster}(x)
}
\arguments{
\item{x}{the object from which to derive the attribute table.}

\item{...}{subset based on logical predicates defined in terms of the
variables in \code{x} or a vector of booleans. Multiple conditions are
combined with &. Only rows where the condition evaluates to TRUE are kept.}
}
\value{
A table of the group attributes of \code{x} or an object where the
  groups table has been subsetted.
}
\description{
Get tabular information of the attributes of groups of features.
}
\examples{
getGroups(x = gtGeoms$polygon)

getGroups(gtRasters$categorical)
}
\seealso{
Other getters: \code{\link{getCRS}},
  \code{\link{getExtent}}, \code{\link{getFeatures}},
  \code{\link{getHistory}}, \code{\link{getPoints}},
  \code{\link{getType}}, \code{\link{getWindow}}
}
\concept{getters}
