% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistory.R
\docType{methods}
\name{getHistory}
\alias{getHistory}
\alias{getHistory,ANY-method}
\alias{getHistory,geom-method}
\alias{getHistory,RasterLayer-method}
\alias{getHistory,RasterBrick-method}
\alias{getHistory,RasterStack-method}
\title{Get the history of a spatial object.}
\usage{
\S4method{getHistory}{ANY}(x)

\S4method{getHistory}{geom}(x)

\S4method{getHistory}{RasterLayer}(x)

\S4method{getHistory}{RasterBrick}(x)

\S4method{getHistory}{RasterStack}(x)
}
\arguments{
\item{x}{the object from which to derive the history.}
}
\value{
A list of the events that lead to \code{x}.
}
\description{
Get the history of a spatial object.
}
\seealso{
Other getters: \code{\link{getCRS}},
  \code{\link{getExtent}}, \code{\link{getPoints}},
  \code{\link{getSubset}}, \code{\link{getTable}},
  \code{\link{getType}}, \code{\link{getWindow}}
}
\concept{getters}
