% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPoints.R
\docType{methods}
\name{getPoints}
\alias{getPoints}
\alias{getPoints,geom-method}
\alias{getPoints,Spatial-method}
\alias{getPoints,sf-method}
\alias{getPoints,ppp-method}
\title{Get the table of coordinates of a spatial object.}
\usage{
\S4method{getPoints}{geom}(x)

\S4method{getPoints}{Spatial}(x)

\S4method{getPoints}{sf}(x)

\S4method{getPoints}{ppp}(x)
}
\arguments{
\item{x}{the object from which to extract the coordinates}
}
\value{
A table of the coordinates \code{x} is made up of.
}
\description{
Get the table of coordinates of a spatial object.
}
\examples{
getPoints(gtGeoms$polygon)

getPoints(gtSP$SpatialPoints)

getPoints(gtSF$multilinestring)

getPoints(gtPPP)
}
\seealso{
Other getters: \code{\link{getCRS}},
  \code{\link{getExtent}}, \code{\link{getHistory}},
  \code{\link{getSubset}}, \code{\link{getTable}},
  \code{\link{getType}}, \code{\link{getWindow}}
}
\concept{getters}
