% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubset.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom-method}
\alias{getSubset,Spatial-method}
\alias{getSubset,sf-method}
\title{Get the subset of a spatial object.}
\usage{
\S4method{getSubset}{geom}(x, ..., slot = "table")

\S4method{getSubset}{Spatial}(x, ...)

\S4method{getSubset}{sf}(x, ...)
}
\arguments{
\item{x}{object to \code{subset}.}

\item{slot}{[\code{character(1)}]\cr the slot in which to determine a subset,
either \code{"table"} or \code{"coords"}.}

\item{subset}{[\code{integerish(.)} | \code{logical(.)} |
\code{character(1)}]\cr coordinates to keep.}
}
\description{
Get the subset of a spatial object.
}
\examples{

library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
(aSfc <- getSubset(nc, attr = "NAME \%in\% c('Ashe', 'Surry')"))
}
