% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubset.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom-method}
\alias{getSubset,Spatial-method}
\alias{getSubset,sf-method}
\title{Get the subset of a spatial object.}
\usage{
\S4method{getSubset}{geom}(x, ..., slot = "table")

\S4method{getSubset}{Spatial}(x, ...)

\S4method{getSubset}{sf}(x, ...)
}
\arguments{
\item{x}{object to \code{subset}.}

\item{...}{Logical predicates defined in terms of the variables in \code{x}.}

\item{slot}{[\code{character(1)}]\cr the slot in which to determine a subset,
either \code{"table"} or \code{"coords"}.}
}
\description{
Get the subset of a spatial object.
}
\examples{
(obj <- gtGeoms$locations) # has column a

# get subset of features that have more than 3 vertices
(getSubset(x = obj, n < 3, slot = "table"))
(obj <- gtSF$polygon) # has column a

(getSubset(x = obj, a == 1))
}
