% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubset.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom-method}
\alias{getSubset,Spatial-method}
\alias{getSubset,sf-method}
\alias{getSubset,ppp-method}
\title{Get the subset of a spatial object.}
\usage{
\S4method{getSubset}{geom}(x, ..., slot = "feat")

\S4method{getSubset}{Spatial}(x, ...)

\S4method{getSubset}{sf}(x, ...)

\S4method{getSubset}{ppp}(x)
}
\arguments{
\item{x}{object to \code{subset}.}

\item{...}{Logical predicates defined in terms of the variables in \code{x}.
Multiple conditions are combined with &. Only rows where the condition
evaluates to TRUE are kept.}

\item{slot}{[\code{character(1)}]\cr the slot in which to determine a subset,
either \code{"vert"} for vertices, \code{"feat"} for features or
\code{"group"} for group tables.}
}
\value{
A subset of \code{x} in its original class.
}
\description{
Get the subset of a spatial object.
}
\examples{
# get the subset of a geom
(obj <- gtGeoms$point)

# the first feature
(getSubset(x = gtGeoms$line, fid == 1))


# get the subset of a Spatial object
(obj <- gtSP$SpatialPolygonsDataFrame)

# the subset where the attribute a equals 2
(getSubset(x = obj, a == 2))


# get the subset of an sf object
(obj <- gtSF$polygon)

# the subset where the attribute a equals 1
(getSubset(x = obj, a == 1))

# getSubset(gtPPP$...)
}
