% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubset.R
\docType{methods}
\name{getSubset}
\alias{getSubset}
\alias{getSubset,geom-method}
\alias{getSubset,Spatial-method}
\alias{getSubset,sf-method}
\title{Get the subset of a spatial object.}
\usage{
\S4method{getSubset}{geom}(x, ..., slot = "table")

\S4method{getSubset}{Spatial}(x, ...)

\S4method{getSubset}{sf}(x, ...)
}
\arguments{
\item{x}{object to \code{subset}.}

\item{...}{Logical predicates defined in terms of the variables in \code{x}.
Multiple conditions are combined with &. Only rows where the condition
evaluates to TRUE are kept.}

\item{slot}{[\code{character(1)}]\cr the slot in which to determine a subset,
either \code{"table"} or \code{"vert"}.}
}
\description{
Get the subset of a spatial object.
}
\examples{
(obj <- gtGeoms$point)

# get feature 1
(getSubset(x = gtGeoms$line, fid == 1))


(obj <- gtSP$SpatialPolygonsDataFrame)
(getSubset(x = obj, a == 2))

(obj <- gtSF$polygon)
(getSubset(x = obj, a == 1))
}
