% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable,geom-method}
\alias{getTable,Spatial-method}
\alias{getTable,sf-method}
\alias{getTable,RasterLayer-method}
\title{Get the attribute table of a spatial object.}
\usage{
\S4method{getTable}{geom}(x)

\S4method{getTable}{Spatial}(x)

\S4method{getTable}{sf}(x)

\S4method{getTable}{RasterLayer}(x)
}
\arguments{
\item{x}{the object from which to derive the attribute table.}
}
\description{
Get the attribute table of a spatial object.
}
\examples{
# the attribute table of ...

# ... a geom
getTable(gtGeoms$mask)

# ... any Spatial* object
getTable(gtSP$SpatialPolygons)

# ... an sf object
getTable(gtSF$multiline)

# ... a RasterLayer
getTable(gtRasters$categorical)
}
