% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable,geom-method}
\alias{getTable,Spatial-method}
\alias{getTable,sf-method}
\alias{getTable,RasterLayer-method}
\alias{getTable,RasterBrick-method}
\alias{getTable,matrix-method}
\title{Get the attribute table of a spatial object.}
\usage{
\S4method{getTable}{geom}(x, slot = NULL)

\S4method{getTable}{Spatial}(x)

\S4method{getTable}{sf}(x)

\S4method{getTable}{RasterLayer}(x)

\S4method{getTable}{RasterBrick}(x)

\S4method{getTable}{matrix}(x)
}
\arguments{
\item{x}{the object from which to derive the attribute table.}

\item{slot}{[\code{character(1)}]\cr One of \code{"vert"}, \code{"feat"} or
\code{"group"} from which to retrieve the attribute table. If left at
\code{NULL}, 'feat' and 'group' will be joined.}
}
\description{
Get the attribute table of a spatial object.
}
\examples{
getTable(gtGeoms$polygon)

getTable(x = gtSP$SpatialPolygons)

getTable(gtSF$multiline)

getTable(gtRasters$categorical)
}
