% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\docType{methods}
\name{getTable}
\alias{getTable}
\alias{getTable,geom-method}
\alias{getTable,Spatial-method}
\alias{getTable,sf-method}
\title{Get the attribute table}
\usage{
\S4method{getTable}{geom}(x)

\S4method{getTable}{Spatial}(x)

\S4method{getTable}{sf}(x)
}
\description{
Get the attribute table of a spatial object.
}
\examples{
# the attribute table of ...

# ... a geom
getTable(rtGeoms$mask)

# ... any Spatial* object
getTable(rtSP$SpatialPolygons)

# ... an sf object
getTable(rtSF$multiline)
}
