% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getType.R
\docType{methods}
\name{getType}
\alias{getType}
\alias{getType,geom-method}
\alias{getType,Spatial-method}
\alias{getType,sf-method}
\alias{getType,ppp-method}
\alias{getType,RasterLayer-method}
\title{Get the type of a spatial object.}
\usage{
\S4method{getType}{geom}(x)

\S4method{getType}{Spatial}(x)

\S4method{getType}{sf}(x)

\S4method{getType}{ppp}(x)

\S4method{getType}{RasterLayer}(x)
}
\arguments{
\item{x}{the object for which to determine the type.}
}
\value{
A vector of two values giving the general type (vector/raster) and
  the specific type/class of \code{x}.
}
\description{
Get the type of a spatial object.
}
\examples{
getType(gtGeoms$polygon)

getType(x = gtSP$SpatialPolygons)

getType(gtSF$multiline)

# getType(gtPPP$...)

getType(gtRasters$categorical)
}
\seealso{
Other getters: \code{\link{getCRS}},
  \code{\link{getExtent}}, \code{\link{getHistory}},
  \code{\link{getPoints}}, \code{\link{getSubset}},
  \code{\link{getTable}}, \code{\link{getWindow}}
}
\concept{getters}
