% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindow.R
\docType{methods}
\name{getWindow}
\alias{getWindow}
\alias{getWindow,geom-method}
\title{Get the reference window of a spatial object.}
\usage{
\S4method{getWindow}{geom}(x)
}
\arguments{
\item{x}{the object from which to derive the reference window.}
}
\description{
Get the reference window of a spatial object.
}
\examples{

getWindow(gtGeoms$line)
}
