% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_tiles.R
\name{gs_tiles}
\alias{gs_tiles}
\title{Create a regular tiling geometry}
\usage{
gs_tiles(window = NULL, cells = NULL, crs = NULL,
  tiling = "rectangular", centroids = FALSE)
}
\arguments{
\item{window}{[\code{data.frame(1)}]\cr the origin (lower left corner) and
the maximum value (upper right corner) of the tiling.}

\item{cells}{[\code{integerish(2)}]\cr number of tiles in \code{x} and
\code{y} dimension.}

\item{crs}{[\code{character(1)}]\cr corrdinate reference system of the
object in proj4 notation.}

\item{tiling}{[\code{character(1)}]\cr pattern of the tiling. Possible
options are \code{"rectangular"} (default), \code{"hexagonal"},
\code{"triangular"}.}

\item{centroids}{[\code{logical(1)}]\cr should the centroids of the tiling be
returned (\code{TRUE}) or should the tiling be returned (\code{FALSE},
default)?}
}
\value{
An invisible \code{geom}.
}
\description{
Create a regular tiling geometry
}
\examples{
# create grid for GFC data
gfcWindow <- data.frame(x = c(-180, 180),
                        y = c(-60, 80))
tiles_gfc <- gs_tiles(window = gfcWindow, cells = c(36, 14), crs = projs$longlat)

# create grid for MODIS data
modWindow <- data.frame(x = c(-20015109.354, 20015109.354),
                        y = c(-10007554.677, 10007554.677))
tiles_modis <- gs_tiles(window = modWindow, cells = c(36, 18), crs = projs$sinu)

# create grid for the sentinel data
#sntWindow <- data.frame(x = c(),
#                        y = c())
#tiles_sentinel <- gs_tiles(window = sntWindow, cells = c(), crs = projs$utm)
}
