% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_geom.R
\name{gt_as_geom}
\alias{gt_as_geom}
\title{Derive a \code{geom} from other spatial objects}
\usage{
gt_as_geom(input)
}
\arguments{
\item{input}{[\code{Spatial*} | \code{sf}]\cr the spatial object to build a
\code{geom} from.}
}
\value{
a \code{geom} of the type that comes closest to the type of the
  input.
}
\description{
See \code{\link{geom-class}} for details on differences between objects of
class \code{geom} and other spatial classes.
}
\details{
bla
}
\examples{
spPoly <- rtSP$SpatialPolygons
plot(spPoly, col = "goldenrod")
visualise(geom = gFrom(input = spPoly))

sfPoly <- rtSF$polygon
plot(sfPoly, col = "goldenrod")
visualise(geom = gFrom(input = sfPoly))

}
