% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_grob.R
\name{gt_as_grob}
\alias{gt_as_grob}
\title{Transform geometry to grob}
\usage{
gt_as_grob(geom)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}
}
\value{
Depending on the provided geometry either a \code{\link{pointsGrob}},
  \code{\link{linesGrob}}, \code{\link{polylineGrob}} or a
  \code{\link{polygonGrob}}.
}
\description{
A \code{\link{grob}} (graphical object) is the grid-package representation of
a \code{geom} and is used for plotting.
}
\examples{
coords <- data.frame(x = c(40, 70, 70, 50, 40, 60, 70, 40, 60,
                           40, 10, 20, 30, 30, 20, 50, 40, 10, 20),
                     y = c(40, 40, 60, 70, 40, 20, 40, 10, 20,
                           40, 20, 20, 50, 40, 40, 70, 40, 20, 60),
                     fid = c(1, 1, 1, 1, 2, 2, 2, 3, 3,
                             3, 3, 3, 4, 4, 4, 5, 5, 5, 5))
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window)

aGrob <- gToGrob(geom = aGeom)
str(aGrob)
}
