% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_as_sp.R
\name{gt_as_sp}
\alias{gt_as_sp}
\title{Transform a \code{geom} to a Spatial* object}
\usage{
gt_as_sp(geom = NULL)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}
}
\value{
Depending on the provided \code{geom} either a
  \code{\link{SpatialPointsDataFrame}}, \code{\link{SpatialLinesDataFrame}}
  or \code{\link{SpatialPolygonsDataFrame}} object.
}
\description{
An object of class \code{Spatial*} is the sp-package representation of a
\code{geom}.
}
\examples{
(spPoints <- gt_as_sp(geom = gtGeoms$point))
(spLines <- gt_as_sp(geom = gtGeoms$line))
(spPolygon <- gt_as_sp(geom = gtGeoms$polygon))
}
