% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_edit.R
\name{gt_edit}
\alias{gt_edit}
\title{Edit the vertices of geometries}
\usage{
gt_edit(panel = NULL, tolerance = 0.01, fid = NULL,
  centroid = FALSE, verbose = FALSE)
}
\arguments{
\item{panel}{[\code{character(1)}]\cr the panel in which to edit vertices
(i.e. the title shown over the plot).}

\item{tolerance}{[\code{numeric(1)}]\cr width of the locator-boxes as
proportion of the overall box width.}

\item{fid}{[\code{integerish(.)}]\cr the features to edit.}

\item{centroid}{[\code{logical(1)}]\cr instead of moving the vertices of a
\code{geom}, move the centroid and thus the overall \code{geom}.}

\item{verbose}{[\code{logical(1)}]\cr be verbose about intermediate steps?}
}
\value{
An invisible \code{geom}.
}
\description{
\code{gt_edit} allows to edit the vertices of any \code{geom}.
}
\examples{
\dontrun{
library(sf)

nc = st_read(system.file("shape/nc.shp", package="sf"))
nc_geom <- gt_sf(input = nc)
visualise(geom = nc_geom, window = data.frame(x = c(-82.1, -81), y = c(36.6, 35.9)))
modGeom <- gt_edit()
}
}
\seealso{
Other geometry tools: \code{\link{gt_group}},
  \code{\link{gt_rotate}}, \code{\link{gt_scale}},
  \code{\link{gt_skew}}, \code{\link{gt_stretch}}
}
\concept{geometry tools}
