% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_group.R
\name{gt_group}
\alias{gt_group}
\title{Group geometries}
\usage{
gt_group(geom, index = NULL, distance = NULL, clusters = NULL, ...)
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{index}{[\code{integerish(.)}]\cr a vector with a value for each vertex,
according to which to group.}

\item{distance}{[\code{numeric(1)}]\cr specific distance of two vertices
below which they will be included in the same cluster; must be within the
range of \code{extent}.}

\item{clusters}{[\code{integerish(1)}]\cr the number of clusters for
\code{\link{kmeans}} clustering.}

\item{...}{[various]\cr additional arguments either to
\code{\link[stats]{hclust}} or to \code{\link[stats]{kmeans}}.}
}
\value{
\code{geom} with grouped coordinates.
}
\description{
\code{gGroup} assigns the vertices of a \code{geom} into groups of features.
}
\details{
Only one of the three arguments \code{index}, \code{distance} or
  \code{clusters} need to be set, as grouping is only carried out by one of
  them.

  In case the geom had an attribute table, this has to be redefined by
  default because it is impossible to determine how these attributes should
  be reattributed without external information.
}
\examples{
coords <- data.frame(x = c(30, 60, 60, 40, 10, 40, 20),
                     y = c(40, 40, 60, 70, 10, 20, 40),
                     fid = c(1, 1, 1, 1, 2, 2, 2))
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window, show = TRUE)

grouped <- gGroup(geom = aGeom, distance = 40)
visualise(geom = grouped)
}
