% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_scale.R
\name{gt_scale}
\alias{gt_scale}
\title{Scale geometries}
\usage{
gt_scale(geom, range = NULL, to = "relative")
}
\arguments{
\item{geom}{[\code{geom}]\cr Object of class \code{\link{geom}}.}

\item{range}{[\code{list(2)}]\cr integerish vector of length two for \code{x}
and \code{y}.}

\item{to}{[\code{character(1)}]\cr the scale to which the coordinates should
be transformed; possible are \code{"relative"} and \code{"absolute"};
ignored in case \code{range != NULL}.}
}
\value{
Scaled \code{geom}.
}
\description{
Relative coordinates are required to work with grobs of \code{geom}s, absolute
coordinates are required to create spatial objects thereof.
}
\examples{
coords <- data.frame(x = c(40, 70, 70, 50),
                     y = c(40, 40, 60, 70),
                     fid = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window, col = "blue")

# change to relative scale and back to absolute
(relCoords <- gScale(geom = aGeom, to = "relative"))
gScale(geom = relCoords, to = "absolute")

gScale(geom = aGeom, range = list(x = c(0, 100), y = c(10, 90)))

}
