% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtGeoms}
\alias{gtGeoms}
\title{Example \code{geom} objects}
\format{The list contains three objects of class geom, a \code{point},
  \code{line} and \code{polygon} object. They are mostly used in the
  example and test-sections of this package.}
\usage{
gtGeoms
}
\description{
A set of three geometries.
}
\keyword{datasets}
