% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeLayout}
\alias{makeLayout}
\title{Make the layout of a plot}
\usage{
makeLayout(x = NULL, window = NULL, theme = gtTheme, image = FALSE,
  ...)
}
\arguments{
\item{x}{the object from which to make the plot.}

\item{window}{[\code{data.frame(1)}] two oposing corners of a rectangle to
which the plot is limited.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}

\item{image}{[\code{logical(1)}]\cr whether or not the raster (brick)
contains an image.}

\item{...}{[]}
}
\description{
Make the layout of a plot
}
