% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeLayout}
\alias{makeLayout}
\title{Make the layout of a plot}
\usage{
makeLayout(x = NULL, theme = gtTheme)
}
\arguments{
\item{x}{[\code{list(.)}]\cr the object, output from \code{makeObject}, from
which to make the plot.}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}
}
\description{
Make the layout of a plot
}
