% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeObject.R
\name{makeObject}
\alias{makeObject}
\title{Make the object to a plot}
\usage{
makeObject(x, window = NULL, image = FALSE, theme = gtTheme, ...)
}
\arguments{
\item{x}{[\code{list(1)}]\cr named list of the object from which to make the
plot.}

\item{window}{[\code{data.frame(1)}] two opposing corners of a rectangle to
which the plot is limited.}

\item{image}{[\code{logical(1)}]\cr whether or not \code{x} is an image}

\item{theme}{[\code{gtTheme(1)}]\cr the theme from which to take graphical
parameters.}

\item{...}{instead of providing a \code{gtTheme}, you can also determine
specific graphic parameters (see \code{\link{gpar}}) separately; see
\code{\link{setTheme}} for details.}
}
\description{
Make the object to a plot
}
