% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaleParameters}
\alias{scaleParameters}
\title{Set the scale of a plot}
\usage{
scaleParameters(attr = NULL, params = NULL)
}
\arguments{
\item{attr}{[\code{data.frame(1)}]\cr the attribute table from which to
derive values.}

\item{params}{[\code{list(7)}]\cr the parameters of the geom that shall be
scaled.}
}
\value{
a list of parameters to a grob.
}
\description{
Set the scale of a plot
}
