% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTable.R
\docType{methods}
\name{setTable}
\alias{setTable}
\alias{setTable,geom-method}
\alias{setTable,Spatial-method}
\alias{setTable,sf-method}
\alias{setTable,sfc-method}
\alias{setTable,RasterLayer-method}
\title{Set the attribute table(s) of a spatial object.}
\usage{
\S4method{setTable}{geom}(x, table = NULL, slot = "feat")

\S4method{setTable}{Spatial}(x, table = NULL)

\S4method{setTable}{sf}(x, table = NULL)

\S4method{setTable}{sfc}(x, table = NULL)

\S4method{setTable}{RasterLayer}(x, table = NULL)
}
\arguments{
\item{x}{the object to which to assign an attribute table.}

\item{table}{[\code{data.frame(.)}]\cr the attribute table.}

\item{slot}{[\code{character(1)}]\cr the slot (of \code{geom}) for which to
set the attribute table, either \code{"vert"}, \code{"feat"} or
\code{"group"}.}
}
\value{
The object \code{x} with an updated attribute table.
}
\description{
Set the attribute table(s) of a spatial object.
}
\details{
If \code{table} does not have columns in common with \code{x}, the
  new columns are simply bound to the original attribute table (if possible).
  If there are common columns, they are joined.
}
\examples{
# set table of a geom
# individual attributes per point/line/polygon feature
getTable(gtGeoms$point)
newAttr <- setTable(x = gtGeoms$point,
                    slot = "vert",
                    table = data.frame(attr = letters[c(1:12)]))
getTable(x = newAttr, slot = "vert")

newAttr2 <- setTable(x = newAttr,
                     slot = "feat",
                     table = data.frame(gid = c(1:3), attr = letters[1:3]))
getTable(x = newAttr2, slot = "feat")
newAttr2

# set table of an sf
sfObj <- gtSF$polygon

# ... with common columns
myAttributes <- data.frame(a = c(2, 1), attr = letters[1:2])
setTable(x = sfObj, table = myAttributes)

# ... without common columns
setTable(x = sfObj, table = myAttributes[2])

# set table to an sfc (no join possible) and transform it to sf thereby
sfcObj = gtSF$polygon$geometry
setTable(x = sfcObj, table = myAttributes)
}
