% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWindow.R
\docType{methods}
\name{setWindow}
\alias{setWindow}
\alias{setWindow,geom-method}
\title{Set the reference window of a spatial object.}
\usage{
\S4method{setWindow}{geom}(x, to)
}
\arguments{
\item{x}{the object in which to set the reference window.}

\item{to}{[\code{data.frame(1)}]\cr two oposing corners.}
}
\description{
Set the reference window of a spatial object.
}
\examples{
visualise(geom = gtGeoms$mask)

buffered <- data.frame(x = c(5074000, 5110000),
                       y = c(4050800, 4076000))
visualise(geom = setWindow(x = gtGeoms$mask,
                           to = buffered))
}
