% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWindow.R
\docType{methods}
\name{setWindow}
\alias{setWindow}
\alias{setWindow,geom-method}
\title{Set the reference window}
\usage{
\S4method{setWindow}{geom}(x, to)
}
\description{
Set the reference window of a spatial object.
}
