% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1geom.R
\name{show,geom-method}
\alias{show,geom-method}
\title{Print geom in the console}
\usage{
\S4method{show}{geom}(object)
}
\arguments{
\item{object}{\link[=geom]{gridded(1)}\cr object to \code{show}.}
}
\description{
Print geom in the console
}
