% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{show,geom-method}
\alias{show,geom-method}
\title{Print geom in the console}
\usage{
\S4method{show}{geom}(object)
}
\arguments{
\item{object}{[\code{geom}]\cr object to \code{show}.}
}
\description{
Print geom in the console
}
