% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise}
\alias{visualise}
\title{Visualise raster and geom objects}
\usage{
visualise(raster = NULL, geom = NULL, window = NULL,
  theme = gtTheme, trace = FALSE, image = FALSE, new = TRUE, ...)
}
\arguments{
\item{raster}{[\code{Raster*} | \code{matrix}]\cr raster object to plot.}

\item{geom}{[\code{geom}]\cr geom to plot.}

\item{window}{[\code{data.frame(1)}]\cr two oposing corners to which the plot
is limited.}

\item{theme}{[\code{list(7)}]\cr Visualising options; see
\code{\link{setTheme}} for details.}

\item{trace}{[\code{logical(1)}]\cr Print the raster object's history (i.e.
the process according to which it has been created) (\code{TRUE}), or
simply plot the object (\code{FALSE}, default).}

\item{image}{[\code{logical(1)}]\cr Does \code{raster} have the channels
\code{red}, \code{green} and \code{blue}, i.e. is it an "image"
(\code{TRUE}) or is this not the case (\code{FALSE}, default)?}

\item{new}{[\code{logical(1)}]\cr force a new plot (\code{TRUE}, default).}

\item{...}{[various]\cr graphical parameters to plot a \code{geom}.}
}
\value{
Returns invisibly an object of class \code{recordedplot}, see
  \code{\link{recordPlot}} for details (and warnings).
}
\description{
Visualise raster and geom objects
}
\details{
To create a plot with your own style, design it with
  \code{\link{setTheme}} and use it in \code{theme}.

  In case you want to plot an image (simiar to
  \code{\link[raster]{plotRGB}}), you have to provide a \code{RasterStack}
  with the three layers \code{red}, \code{green} and \code{blue} and set
  \code{image = TRUE}.
}
\examples{
# define a geometry
coords <- data.frame(x = c(30, 60, 60, 40),
                     y = c(40, 40, 60, 70),
                     fid = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
(aGeom <- gs_polygon(anchor = coords, window = window, col = "blue"))

# if plotted on top of an existing plot, the relative coordinate values
# will be used to construct the grob.
visualise(raster = gtRasters$continuous, geom = aGeom)
visualise(geom = aGeom)

}
